Refactor gui module into focused submodules
