+++ src/gui/mod.rs
@@ -1,3 +1,4 @@
+mod actions;
 mod add_action_dialog;
 mod add_bookmark_dialog;
 mod alias_dialog;
@@ -19,9 +20,12 @@ mod mouse_gestures_dialog;
 mod note_graph_dialog;
 mod note_panel;
 mod notes_dialog;
+mod render;
 mod screenshot_editor;
+mod search;
 mod shell_cmd_dialog;
 mod snippet_dialog;
+mod state;
 mod tempfile_alias_dialog;
 mod tempfile_dialog;
 mod theme_settings_dialog;
@@ -32,6 +36,7 @@ mod todo_view_dialog;
 mod unused_assets_dialog;
 pub(crate) mod volume_data;
 mod volume_dialog;
+mod watch;
 
 pub use add_action_dialog::AddActionDialog;
 pub use add_bookmark_dialog::AddBookmarkDialog;
@@ -106,6 +111,7 @@ use fuzzy_matcher::skim::SkimMatcherV2;
 use fuzzy_matcher::FuzzyMatcher;
 use notify::{Config, EventKind, RecommendedWatcher, RecursiveMode, Watcher};
 use once_cell::sync::Lazy;
+use search::{COMPLETION_REBUILD_DEBOUNCE, NOTE_SEARCH_DEBOUNCE};
 use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
 use std::fmt::Display;
@@ -118,6 +124,12 @@ use std::sync::{
 };
 use std::time::{Duration, Instant};
 use url::Url;
+use watch::watch_file;
+
+pub(crate) use state::{
+    ActivationSource, PendingConfirmAction, ResultContextMenuKind, TestWatchEvent, UiErrorEvent,
+    WatchEvent,
+};
 
 const SUBCOMMANDS: &[&str] = &[
     "add", "rm", "list", "clear", "open", "new", "alias", "set", "pause", "resume", "cancel",
@@ -126,9 +138,6 @@ const SUBCOMMANDS: &[&str] = &[
 
 /// Prefix used to search user saved applications.
 pub const APP_PREFIX: &str = "app";
-const NOTE_SEARCH_DEBOUNCE: Duration = Duration::from_secs(1);
-const COMPLETION_REBUILD_DEBOUNCE: Duration = Duration::from_millis(120);
-
 fn scale_ui<R>(ui: &mut egui::Ui, scale: f32, add_contents: impl FnOnce(&mut egui::Ui) -> R) -> R {
     ui.scope(|ui| {
         if (scale - 1.0).abs() > f32::EPSILON {
@@ -143,131 +152,6 @@ fn scale_ui<R>(ui: &mut egui::Ui, scale: f32, add_contents: impl FnOnce(&mut egu
         add_contents(ui)
     })
     .inner
-}
-
-#[derive(Clone)]
-pub enum WatchEvent {
-    Actions,
-    Folders,
-    Bookmarks,
-    Clipboard,
-    Snippets,
-    Notes,
-    Todos,
-    Favorites,
-    Gestures,
-    Dashboard(DashboardEvent),
-    Recycle(Result<(), String>),
-    ExecuteAction(Action),
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ActivationSource {
-    Enter,
-    Click,
-    Dashboard,
-    Gesture,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct UiErrorEvent {
-    pub context: &'static str,
-    pub message: String,
-}
-
-impl UiErrorEvent {
-    pub fn new(context: &'static str, err: impl Display) -> Self {
-        Self {
-            context,
-            message: err.to_string(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum ResultContextMenuKind {
-    Folder,
-    Bookmark,
-    Timer { id: u64 },
-    Stopwatch { id: u64 },
-    Snippet,
-    Tempfile,
-    Note { slug: String },
-    Clipboard { idx: usize, label: String },
-    Todo { idx: usize },
-    Default,
-}
-
-impl ActivationSource {
-    fn label(self) -> &'static str {
-        match self {
-            Self::Enter => "enter",
-            Self::Click => "click",
-            Self::Dashboard => "dashboard",
-            Self::Gesture => "gesture",
-        }
-    }
-}
-
-#[derive(Clone)]
-struct PendingConfirmAction {
-    action: Action,
-    query_override: Option<String>,
-    source: ActivationSource,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum TestWatchEvent {
-    Actions,
-    Folders,
-    Bookmarks,
-}
-
-impl From<WatchEvent> for TestWatchEvent {
-    fn from(value: WatchEvent) -> Self {
-        match value {
-            WatchEvent::Actions => TestWatchEvent::Actions,
-            WatchEvent::Folders => TestWatchEvent::Folders,
-            WatchEvent::Bookmarks => TestWatchEvent::Bookmarks,
-            WatchEvent::Clipboard => TestWatchEvent::Actions,
-            WatchEvent::Snippets => TestWatchEvent::Actions,
-            WatchEvent::Notes => TestWatchEvent::Actions,
-            WatchEvent::Todos => TestWatchEvent::Actions,
-            WatchEvent::Favorites => TestWatchEvent::Actions,
-            WatchEvent::Gestures => TestWatchEvent::Actions,
-            WatchEvent::Dashboard(_) => TestWatchEvent::Actions,
-            WatchEvent::Recycle(_) => unreachable!(),
-            WatchEvent::ExecuteAction(_) => TestWatchEvent::Actions,
-        }
-    }
-}
-
-fn watch_file(
-    path: &Path,
-    tx: Sender<WatchEvent>,
-    event: WatchEvent,
-) -> notify::Result<RecommendedWatcher> {
-    let mut watcher = RecommendedWatcher::new(
-        move |res: notify::Result<notify::Event>| match res {
-            Ok(ev) => {
-                if matches!(
-                    ev.kind,
-                    EventKind::Modify(_) | EventKind::Create(_) | EventKind::Remove(_)
-                ) {
-                    let _ = tx.send(event.clone());
-                }
-            }
-            Err(e) => tracing::error!("watch error: {:?}", e),
-        },
-        Config::default(),
-    )?;
-    watcher
-        .watch(path, RecursiveMode::NonRecursive)
-        .or_else(|_| {
-            let parent = path.parent().unwrap_or_else(|| Path::new("."));
-            watcher.watch(parent, RecursiveMode::NonRecursive)
-        })?;
-    Ok(watcher)
 }
 
 fn push_toast(toasts: &mut Toasts, toast: Toast) {
@@ -614,358 +498,6 @@ impl CachedSearchEntry {
 }
 
 impl LauncherApp {
-    fn normalize_alias(alias: Option<String>) -> (Option<String>, Option<String>) {
-        let alias_lc = alias.as_ref().map(|text| text.to_lowercase());
-        (alias, alias_lc)
-    }
-
-    fn folder_alias_maps() -> (
-        HashMap<String, Option<String>>,
-        HashMap<String, Option<String>>,
-    ) {
-        let mut aliases = HashMap::new();
-        let mut aliases_lc = HashMap::new();
-        for folder in crate::plugins::folders::load_folders(crate::plugins::folders::FOLDERS_FILE)
-            .unwrap_or_else(|_| crate::plugins::folders::default_folders())
-        {
-            let (alias, alias_lc) = Self::normalize_alias(folder.alias);
-            aliases.insert(folder.path.clone(), alias);
-            aliases_lc.insert(folder.path, alias_lc);
-        }
-        (aliases, aliases_lc)
-    }
-
-    fn bookmark_alias_maps() -> (
-        HashMap<String, Option<String>>,
-        HashMap<String, Option<String>>,
-    ) {
-        let mut aliases = HashMap::new();
-        let mut aliases_lc = HashMap::new();
-        for bookmark in
-            crate::plugins::bookmarks::load_bookmarks(crate::plugins::bookmarks::BOOKMARKS_FILE)
-                .unwrap_or_default()
-        {
-            let (alias, alias_lc) = Self::normalize_alias(bookmark.alias);
-            aliases.insert(bookmark.url.clone(), alias);
-            aliases_lc.insert(bookmark.url, alias_lc);
-        }
-        (aliases, aliases_lc)
-    }
-
-    fn alias_matches_lc(&self, action: &str, query_lc: &str) -> bool {
-        self.folder_aliases_lc
-            .get(action)
-            .or_else(|| self.bookmark_aliases_lc.get(action))
-            .and_then(|v| v.as_ref())
-            .map(|s| s.contains(query_lc))
-            .unwrap_or(false)
-    }
-
-    fn is_exact_match_mode(&self) -> bool {
-        // `match_exact` is a strict override: if enabled, we always bypass fuzzy scoring.
-        self.match_exact || self.fuzzy_weight <= 0.0
-    }
-
-    fn matches_exact_display_text(cached: &CachedSearchEntry, query_lc: &str) -> bool {
-        let query_lc = query_lc.trim();
-        if query_lc.is_empty() {
-            return true;
-        }
-        cached.label_lc.contains(query_lc)
-    }
-
-    fn should_bypass_exact_post_filter(query: &str, action: &str) -> bool {
-        // `query:*` actions are command suggestions that should still participate in
-        // exact display-text filtering when users are browsing command names/options.
-        if action.starts_with("query:") {
-            return false;
-        }
-
-        let mut parts = query.split_whitespace();
-        let Some(head) = parts.next().map(str::to_ascii_lowercase) else {
-            return false;
-        };
-        let Some(subcommand) = parts.next().map(str::to_ascii_lowercase) else {
-            return false;
-        };
-
-        // Only bypass launcher-side exact display filtering when the query is an
-        // explicit plugin command whose plugin already returned resolved outputs.
-        // Example: `note today` / `note search <term>` yielding `note:new:*` or
-        // `note:open:*` actions; re-filtering those by label text can hide valid results.
-        matches!(head.as_str(), "note" | "notes")
-            && matches!(
-                subcommand.as_str(),
-                "today"
-                    | "search"
-                    | "links"
-                    | "link"
-                    | "list"
-                    | "open"
-                    | "new"
-                    | "add"
-                    | "create"
-                    | "graph"
-                    | "templates"
-                    | "tag"
-                    | "rm"
-            )
-            && action.starts_with("note:")
-    }
-
-    fn has_diagnostics_widget(&self) -> bool {
-        self.dashboard
-            .slots
-            .iter()
-            .any(|slot| slot.widget == "diagnostics")
-    }
-
-    pub fn update_action_cache(&mut self) {
-        self.action_cache = self
-            .actions
-            .iter()
-            .map(CachedSearchEntry::from_action)
-            .collect();
-        self.action_filter_metadata = self
-            .actions
-            .iter()
-            .map(ActionFilterMetadata::from_action)
-            .collect();
-        self.actions_by_id = self
-            .actions
-            .iter()
-            .map(|a| (a.action.clone(), a.clone()))
-            .collect();
-        self.action_completion_dirty = true;
-        self.schedule_completion_rebuild();
-    }
-
-    pub fn update_command_cache(&mut self) {
-        let mut cmds = self
-            .plugins
-            .commands_filtered(self.enabled_plugins.as_ref());
-        cmds.sort_by_cached_key(|a| a.label.to_lowercase());
-        self.command_search_cache = cmds.iter().map(CachedSearchEntry::from_action).collect();
-        self.command_cache = cmds;
-        self.command_completion_dirty = true;
-        self.schedule_completion_rebuild();
-    }
-
-    fn schedule_completion_rebuild(&mut self) {
-        self.completion_rebuild_after = Some(Instant::now() + COMPLETION_REBUILD_DEBOUNCE);
-        self.completion_index = None;
-        self.autocomplete_index = 0;
-        self.suggestions.clear();
-    }
-
-    fn maybe_rebuild_completion_index(&mut self, now: Instant) {
-        let should_rebuild = self
-            .completion_rebuild_after
-            .is_some_and(|scheduled| now >= scheduled)
-            && (self.action_completion_dirty || self.command_completion_dirty);
-        if should_rebuild {
-            self.update_completion_index();
-            self.action_completion_dirty = false;
-            self.command_completion_dirty = false;
-            self.completion_rebuild_after = None;
-        }
-    }
-
-    fn rebuild_completion_index_now(&mut self) {
-        if self.action_completion_dirty || self.command_completion_dirty {
-            self.update_completion_index();
-            self.action_completion_dirty = false;
-            self.command_completion_dirty = false;
-        }
-        self.completion_rebuild_after = None;
-    }
-
-    pub fn process_watch_events(&mut self) {
-        while let Ok(ev) = self.rx.try_recv() {
-            match ev {
-                WatchEvent::Actions => {
-                    if let Ok(mut acts) = load_actions(&self.actions_path) {
-                        let custom_len = acts.len();
-                        self.custom_len = custom_len;
-                        if let Some(paths) = &self.index_paths {
-                            let options =
-                                indexer::IndexOptions::with_max_items(self.max_indexed_items);
-                            for batch in indexer::index_paths_batched(paths, options) {
-                                match batch {
-                                    Ok(idx) => {
-                                        acts.extend(idx);
-                                        self.actions = Arc::new(acts.clone());
-                                        self.update_action_cache();
-                                        self.search();
-                                    }
-                                    Err(e) => {
-                                        tracing::error!(error = %e, "failed to index paths");
-                                        self.report_error_message(
-                                            "launcher",
-                                            format!("Failed to index paths: {e}"),
-                                        );
-                                        break;
-                                    }
-                                }
-                            }
-                        }
-                        self.actions = Arc::new(acts);
-                        self.update_action_cache();
-                        self.search();
-                        crate::actions::bump_actions_version();
-                        tracing::info!("actions reloaded");
-                    }
-                }
-                WatchEvent::Folders => {
-                    let (aliases, aliases_lc) = Self::folder_alias_maps();
-                    self.folder_aliases = aliases;
-                    self.folder_aliases_lc = aliases_lc;
-                    self.search();
-                }
-                WatchEvent::Bookmarks => {
-                    let (aliases, aliases_lc) = Self::bookmark_alias_maps();
-                    self.bookmark_aliases = aliases;
-                    self.bookmark_aliases_lc = aliases_lc;
-                    self.search();
-                }
-                WatchEvent::Clipboard => {
-                    self.dashboard_data_cache.refresh_clipboard();
-                }
-                WatchEvent::Snippets => {
-                    self.dashboard_data_cache.refresh_snippets();
-                }
-                WatchEvent::Notes => {
-                    self.dashboard_data_cache.refresh_notes();
-                }
-                WatchEvent::Todos => {
-                    self.dashboard_data_cache.refresh_todos();
-                }
-                WatchEvent::Favorites => {
-                    self.dashboard_data_cache.refresh_favorites();
-                }
-                WatchEvent::Gestures => {
-                    self.dashboard_data_cache.refresh_gestures();
-                }
-                WatchEvent::Dashboard(_) => {
-                    self.dashboard.reload();
-                    for warn in &self.dashboard.warnings {
-                        tracing::warn!("dashboard: {}", warn);
-                        if self.enable_toasts {
-                            push_toast(
-                                &mut self.toasts,
-                                Toast {
-                                    text: warn.clone().into(),
-                                    kind: ToastKind::Warning,
-                                    options: ToastOptions::default()
-                                        .duration_in_seconds(self.toast_duration as f64),
-                                },
-                            );
-                        }
-                    }
-                }
-                WatchEvent::Recycle(res) => match res {
-                    Ok(()) => {
-                        if self.enable_toasts {
-                            push_toast(
-                                &mut self.toasts,
-                                Toast {
-                                    text: "Emptied Recycle Bin".into(),
-                                    kind: ToastKind::Success,
-                                    options: ToastOptions::default()
-                                        .duration_in_seconds(self.toast_duration as f64),
-                                },
-                            );
-                        }
-                    }
-                    Err(e) => {
-                        let msg = format!("Failed to empty recycle bin: {e}");
-                        self.report_error_message("recycle.empty", msg);
-                    }
-                },
-                WatchEvent::ExecuteAction(action) => {
-                    self.activate_action(action, None, ActivationSource::Gesture);
-                }
-            }
-        }
-        self.maybe_rebuild_completion_index(Instant::now());
-    }
-
-    fn update_completion_index(&mut self) {
-        let mut entries: Vec<String> = Vec::new();
-        entries.extend(self.command_cache.iter().map(|a| a.label.to_lowercase()));
-        for a in self.actions.iter() {
-            entries.push(format!("app {}", a.label.to_lowercase()));
-        }
-        entries.sort();
-        entries.dedup();
-        let mut builder = MapBuilder::memory();
-        for (i, k) in entries.iter().enumerate() {
-            if let Err(e) = builder.insert(k, i as u64) {
-                tracing::warn!(key = %k, ?e, "failed to insert key into completion index");
-            }
-        }
-        let map = Map::new(builder.into_inner().unwrap()).unwrap();
-        self.completion_index = Some(map);
-        self.update_suggestions();
-    }
-
-    fn update_suggestions(&mut self) {
-        self.autocomplete_index = 0;
-        self.suggestions.clear();
-        if !self.query_autocomplete
-            || self.query.is_empty()
-            || self.should_show_dashboard(self.query.as_str())
-        {
-            return;
-        }
-        if let Some(ref index) = self.completion_index {
-            let q = self.query.to_lowercase();
-            let mut stream = index.range().ge(q.as_str()).into_stream();
-            while let Some((k, _)) = stream.next() {
-                let key = std::str::from_utf8(k).unwrap();
-                if !key.starts_with(&q) {
-                    break;
-                }
-                if key != q {
-                    self.suggestions.push(key.to_string());
-                }
-                if self.suggestions.len() >= 5 {
-                    break;
-                }
-            }
-        }
-    }
-
-    fn is_note_search_query(query: &str) -> bool {
-        query.trim_start().to_lowercase().starts_with("note search")
-    }
-
-    fn note_search_debounce_ready(
-        last_change: Option<Instant>,
-        now: Instant,
-        debounce: Duration,
-    ) -> bool {
-        last_change
-            .map(|changed_at| now.duration_since(changed_at) >= debounce)
-            .unwrap_or(false)
-    }
-
-    fn maybe_run_note_search_debounce(&mut self) {
-        if !Self::is_note_search_query(&self.query) {
-            self.last_note_search_change = None;
-            return;
-        }
-
-        if Self::note_search_debounce_ready(
-            self.last_note_search_change,
-            Instant::now(),
-            NOTE_SEARCH_DEBOUNCE,
-        ) {
-            self.search();
-            self.last_note_search_change = None;
-        }
-    }
-
     pub fn plugin_enabled(&self, name: &str) -> bool {
         match &self.enabled_plugins {
             Some(set) => set.contains(name),
@@ -1713,252 +1245,6 @@ impl LauncherApp {
         app
     }
 
-    pub fn search(&mut self) {
-        if self.last_results_valid && self.query == self.last_search_query {
-            self.selected = None;
-            return;
-        }
-
-        let trimmed = self.query.trim();
-        let trimmed_lc = trimmed.to_lowercase();
-        self.last_timer_query =
-            trimmed.starts_with("timer list") || trimmed.starts_with("alarm list");
-        self.last_stopwatch_query = trimmed.starts_with("sw list");
-        if trimmed.is_empty() {
-            self.autocomplete_index = 0;
-            self.suggestions.clear();
-            let mut res = self.command_cache.clone();
-            for a in self.actions.iter() {
-                res.push(Action {
-                    label: format!("app {}", a.label),
-                    desc: a.desc.clone(),
-                    action: a.action.clone(),
-                    args: a.args.clone(),
-                });
-            }
-            self.results = res;
-            self.selected = None;
-            self.recompute_query_results_layout();
-            return;
-        }
-
-        let mut res: Vec<(Action, f32)> = Vec::new();
-
-        let search_actions =
-            trimmed_lc == APP_PREFIX || trimmed_lc.starts_with(&format!("{} ", APP_PREFIX));
-        let action_query = if search_actions {
-            if trimmed_lc == APP_PREFIX {
-                "".to_string()
-            } else {
-                trimmed.splitn(2, ' ').nth(1).unwrap_or("").to_string()
-            }
-        } else {
-            String::new()
-        };
-        let action_query_lc = action_query.to_lowercase();
-
-        if trimmed_lc.starts_with("g ") {
-            res.extend(self.search_plugins(trimmed, &trimmed_lc));
-        } else {
-            if search_actions {
-                res.extend(self.search_actions(&action_query, &action_query_lc));
-            }
-            res.extend(self.search_plugins(trimmed, &trimmed_lc));
-        }
-
-        self.apply_usage_weight(&mut res);
-
-        res.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
-
-        self.results = res.into_iter().map(|(a, _)| a).collect();
-        self.selected = None;
-        self.last_search_query = self.query.clone();
-        self.last_results_valid = true;
-        self.update_suggestions();
-        self.recompute_query_results_layout();
-    }
-
-    fn search_actions(&self, query: &str, _query_lc: &str) -> Vec<(Action, f32)> {
-        let (filtered_query, filters) = split_action_filters(query);
-        let filtered_query = filtered_query.trim();
-        let filtered_query_lc = filtered_query.to_lowercase();
-        let query = filtered_query;
-        let query_lc = filtered_query_lc.as_str();
-
-        let mut res = Vec::new();
-        if query.is_empty() {
-            for (i, a) in self.actions.iter().enumerate() {
-                if action_matches_filters(&self.action_filter_metadata[i], &filters) {
-                    res.push((a.clone(), 0.0));
-                }
-            }
-        } else {
-            for (i, a) in self.actions.iter().enumerate() {
-                if !action_matches_filters(&self.action_filter_metadata[i], &filters) {
-                    continue;
-                }
-
-                let cached = &self.action_cache[i];
-                if self.is_exact_match_mode() {
-                    let alias_match = self.alias_matches_lc(&a.action, query_lc);
-                    let label_match = Self::matches_exact_display_text(cached, query_lc);
-                    // Prefer displayed label text, but keep `desc`/aliases as supplemental
-                    // filters for compatibility with existing query behavior.
-                    let desc_match = cached.desc_lc.contains(query_lc);
-                    let action_match = cached.action_lc.contains(query_lc);
-                    if label_match || desc_match || action_match || alias_match {
-                        let score = if alias_match { 1.0 } else { 0.0 };
-                        res.push((a.clone(), score));
-                    }
-                } else {
-                    let s1 = self.matcher.fuzzy_match(&a.label, query);
-                    let s2 = self.matcher.fuzzy_match(&a.desc, query);
-                    if let Some(score) = s1.max(s2) {
-                        res.push((a.clone(), score as f32 * self.fuzzy_weight));
-                    }
-                }
-            }
-        }
-        res
-    }
-
-    fn search_plugins(&self, trimmed: &str, trimmed_lc: &str) -> Vec<(Action, f32)> {
-        let mut res = Vec::new();
-        if trimmed_lc.starts_with("g ") {
-            let filter = std::collections::HashSet::from(["web_search".to_string()]);
-            let plugin_results = self.plugins.search_filtered(
-                &self.query,
-                Some(&filter),
-                self.enabled_capabilities.as_ref(),
-            );
-            let query_term = trimmed_lc.splitn(2, ' ').nth(1).unwrap_or("");
-            for a in plugin_results {
-                let cached = CachedSearchEntry::from_action(&a);
-                if self.is_exact_match_mode() {
-                    if Self::should_bypass_exact_post_filter(trimmed, &a.action) {
-                        // Plugin commands like `note today`/`note search <term>` already
-                        // returned concrete results (e.g. `note:new:*`, `note:open:*`).
-                        // Re-filtering by label/desc text can hide valid plugin-resolved
-                        // outputs, so keep them as-is in exact mode.
-                        res.push((a, 0.0));
-                        continue;
-                    }
-                    if query_term.is_empty() {
-                        res.push((a, 0.0));
-                    } else {
-                        let alias_match = self.alias_matches_lc(&a.action, query_term);
-                        let label_match = Self::matches_exact_display_text(&cached, query_term);
-                        let desc_match = cached.desc_lc.contains(query_term);
-                        let action_match = cached.action_lc.contains(query_term);
-                        if label_match || desc_match || action_match || alias_match {
-                            let score = if alias_match { 1.0 } else { 0.0 };
-                            res.push((a, score));
-                        }
-                    }
-                } else {
-                    let score = if self.query.is_empty() {
-                        0.0
-                    } else {
-                        self.matcher
-                            .fuzzy_match(&a.label, &self.query)
-                            .max(self.matcher.fuzzy_match(&a.desc, &self.query))
-                            .unwrap_or(0) as f32
-                            * self.fuzzy_weight
-                    };
-                    res.push((a, score));
-                }
-            }
-            return res;
-        }
-
-        let plugin_results = self.plugins.search_filtered(
-            &self.query,
-            self.enabled_plugins.as_ref(),
-            self.enabled_capabilities.as_ref(),
-        );
-
-        if plugin_results.is_empty() && !trimmed.is_empty() {
-            for (a, cached) in self
-                .command_cache
-                .iter()
-                .zip(self.command_search_cache.iter())
-            {
-                if self.is_exact_match_mode() {
-                    let alias_match = self.alias_matches_lc(&a.action, trimmed_lc);
-                    let label_match = Self::matches_exact_display_text(cached, trimmed_lc);
-                    let desc_match = cached.desc_lc.contains(trimmed_lc);
-                    let action_match = cached.action_lc.contains(trimmed_lc);
-                    if label_match || desc_match || action_match || alias_match {
-                        let score = if alias_match { 1.0 } else { 0.0 };
-                        res.push((a.clone(), score));
-                    }
-                } else {
-                    let s1 = self.matcher.fuzzy_match(&a.label, trimmed);
-                    let s2 = self.matcher.fuzzy_match(&a.desc, trimmed);
-                    if let Some(score) = s1.max(s2) {
-                        res.push((a.clone(), score as f32 * self.fuzzy_weight));
-                    }
-                }
-            }
-        } else {
-            let tail = trimmed_lc.splitn(2, " ").nth(1).unwrap_or("");
-            let mut query_term = tail.splitn(3, " ").nth(1).unwrap_or("").to_string();
-            if query_term.is_empty() {
-                let parts: Vec<&str> = tail.split_whitespace().collect();
-                if parts.len() == 1 && !SUBCOMMANDS.contains(&parts[0]) {
-                    query_term = parts[0].to_string();
-                } else if parts.len() > 1 {
-                    query_term = parts[1..].join(" ");
-                }
-            }
-            let query_term_lc = query_term.to_lowercase();
-            for a in plugin_results {
-                let cached = CachedSearchEntry::from_action(&a);
-                if self.is_exact_match_mode() {
-                    if Self::should_bypass_exact_post_filter(trimmed, &a.action) {
-                        // Explicit plugin commands can resolve into result lists/artifacts.
-                        // Preserve those resolved actions in exact mode instead of applying
-                        // a second label/description exact filter in the launcher layer.
-                        res.push((a, 0.0));
-                        continue;
-                    }
-                    if query_term_lc.is_empty() {
-                        res.push((a, 0.0));
-                    } else {
-                        let alias_match = self.alias_matches_lc(&a.action, &query_term_lc);
-                        let label_match = Self::matches_exact_display_text(&cached, &query_term_lc);
-                        let desc_match = cached.desc_lc.contains(&query_term_lc);
-                        let action_match = cached.action_lc.contains(&query_term_lc);
-                        if label_match || desc_match || action_match || alias_match {
-                            let score = if alias_match { 1.0 } else { 0.0 };
-                            res.push((a, score));
-                        }
-                    }
-                } else {
-                    let score = if self.query.is_empty() {
-                        0.0
-                    } else {
-                        self.matcher
-                            .fuzzy_match(&a.label, &self.query)
-                            .max(self.matcher.fuzzy_match(&a.desc, &self.query))
-                            .unwrap_or(0) as f32
-                            * self.fuzzy_weight
-                    };
-                    res.push((a, score));
-                }
-            }
-        }
-
-        res
-    }
-
-    fn apply_usage_weight(&self, res: &mut Vec<(Action, f32)>) {
-        for (a, score) in res.iter_mut() {
-            *score += self.usage.get(&a.action).cloned().unwrap_or(0) as f32 * self.usage_weight;
-        }
-    }
-
-    #[cfg_attr(test, allow(dead_code))]
     pub fn maybe_refresh_timer_list(&mut self) {
         let trimmed = self.query.trim();
         if (trimmed.starts_with("timer list")
@@ -2323,623 +1609,6 @@ impl LauncherApp {
         None
     }
 
-    fn result_context_menu_kind(&self, action: &Action) -> ResultContextMenuKind {
-        if self.folder_aliases.contains_key(&action.action) && !action.action.starts_with("folder:")
-        {
-            ResultContextMenuKind::Folder
-        } else if self.bookmark_aliases.contains_key(&action.action) {
-            ResultContextMenuKind::Bookmark
-        } else if action.desc == "Timer" && action.action.starts_with("timer:show:") {
-            action.action[11..]
-                .parse::<u64>()
-                .map(|id| ResultContextMenuKind::Timer { id })
-                .unwrap_or(ResultContextMenuKind::Default)
-        } else if action.desc == "Stopwatch" && action.action.starts_with("stopwatch:show:") {
-            action.action["stopwatch:show:".len()..]
-                .parse::<u64>()
-                .map(|id| ResultContextMenuKind::Stopwatch { id })
-                .unwrap_or(ResultContextMenuKind::Default)
-        } else if action.desc == "Snippet" {
-            ResultContextMenuKind::Snippet
-        } else if action.desc == "Tempfile" && !action.action.starts_with("tempfile:") {
-            ResultContextMenuKind::Tempfile
-        } else if action.desc == "Note" && action.action.starts_with("note:open:") {
-            let slug = action.action.rsplit(':').next().unwrap_or("").to_string();
-            ResultContextMenuKind::Note { slug }
-        } else if action.desc == "Clipboard" && action.action.starts_with("clipboard:copy:") {
-            if let Ok(idx) = action
-                .action
-                .rsplit(':')
-                .next()
-                .unwrap_or("")
-                .parse::<usize>()
-            {
-                ResultContextMenuKind::Clipboard {
-                    idx,
-                    label: action.label.clone(),
-                }
-            } else {
-                ResultContextMenuKind::Default
-            }
-        } else if action.desc == "Todo" && action.action.starts_with("todo:done:") {
-            action
-                .action
-                .rsplit(':')
-                .next()
-                .unwrap_or("")
-                .parse::<usize>()
-                .map(|idx| ResultContextMenuKind::Todo { idx })
-                .unwrap_or(ResultContextMenuKind::Default)
-        } else {
-            ResultContextMenuKind::Default
-        }
-    }
-
-    fn attach_result_context_menu(
-        &mut self,
-        action: &Action,
-        menu_resp: egui::Response,
-        refresh: &mut bool,
-        set_focus: &mut bool,
-    ) -> egui::Response {
-        let custom_idx = self
-            .actions
-            .iter()
-            .take(self.custom_len)
-            .position(|act| act.action == action.action && act.label == action.label);
-        let query = self.query.trim().to_string();
-
-        match self.result_context_menu_kind(action) {
-            ResultContextMenuKind::Folder => {
-                menu_resp.clone().context_menu(|ui| {
-                    if ui.button("Set Alias").clicked() {
-                        self.alias_dialog.open(&action.action);
-                        ui.close_menu();
-                    }
-                    if ui.button("Remove Folder").clicked() {
-                        if let Err(e) = crate::plugins::folders::remove_folder(
-                            crate::plugins::folders::FOLDERS_FILE,
-                            &action.action,
-                        ) {
-                            self.report_error_message(
-                                "launcher",
-                                format!("Failed to remove folder: {e}"),
-                            );
-                        } else {
-                            *refresh = true;
-                            *set_focus = true;
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Removed folder {}", action.label).into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if let Some(idx_act) = custom_idx {
-                        if ui.button("Edit App").clicked() {
-                            self.editor.open_edit(idx_act, &self.actions[idx_act]);
-                            self.show_editor = true;
-                            ui.close_menu();
-                        }
-                    }
-                    self.pin_result_menu(ui, action);
-                });
-            }
-            ResultContextMenuKind::Bookmark => {
-                menu_resp.clone().context_menu(|ui| {
-                    if ui.button("Set Alias").clicked() {
-                        self.bookmark_alias_dialog.open(&action.action);
-                        ui.close_menu();
-                    }
-                    if ui.button("Remove Bookmark").clicked() {
-                        if let Err(e) = crate::plugins::bookmarks::remove_bookmark(
-                            crate::plugins::bookmarks::BOOKMARKS_FILE,
-                            &action.action,
-                        ) {
-                            self.report_error_message(
-                                "launcher",
-                                format!("Failed to remove bookmark: {e}"),
-                            );
-                        } else {
-                            *refresh = true;
-                            *set_focus = true;
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Removed bookmark {}", action.label).into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if let Some(idx_act) = custom_idx {
-                        if ui.button("Edit App").clicked() {
-                            self.editor.open_edit(idx_act, &self.actions[idx_act]);
-                            self.show_editor = true;
-                            ui.close_menu();
-                        }
-                    }
-                    self.pin_result_menu(ui, action);
-                });
-            }
-            ResultContextMenuKind::Timer { id } => {
-                menu_resp.clone().context_menu(|ui| {
-                    if ui.button("Pause Timer").clicked() {
-                        crate::plugins::timer::pause_timer(id);
-                        if query.starts_with("timer list") {
-                            *refresh = true;
-                            *set_focus = true;
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Paused timer {}", action.label).into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if ui.button("Remove Timer").clicked() {
-                        crate::plugins::timer::cancel_timer(id);
-                        if query.starts_with("timer list") {
-                            *refresh = true;
-                            *set_focus = true;
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Removed timer {}", action.label).into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if let Some(idx_act) = custom_idx {
-                        if ui.button("Edit App").clicked() {
-                            self.editor.open_edit(idx_act, &self.actions[idx_act]);
-                            self.show_editor = true;
-                            ui.close_menu();
-                        }
-                    }
-                    self.pin_result_menu(ui, action);
-                });
-            }
-            ResultContextMenuKind::Stopwatch { id } => {
-                menu_resp.clone().context_menu(|ui| {
-                    if ui.button("Pause Stopwatch").clicked() {
-                        crate::plugins::stopwatch::pause_stopwatch(id);
-                        if query.starts_with("sw list") {
-                            *refresh = true;
-                            *set_focus = true;
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Paused stopwatch {}", action.label).into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if ui.button("Resume Stopwatch").clicked() {
-                        crate::plugins::stopwatch::resume_stopwatch(id);
-                        if query.starts_with("sw list") {
-                            *refresh = true;
-                            *set_focus = true;
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Resumed stopwatch {}", action.label).into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if ui.button("Stop Stopwatch").clicked() {
-                        crate::plugins::stopwatch::stop_stopwatch(id);
-                        if query.starts_with("sw list") {
-                            *refresh = true;
-                            *set_focus = true;
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Stopped stopwatch {}", action.label).into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if ui.button("Copy Time").clicked() {
-                        if let Some(time) = crate::plugins::stopwatch::format_elapsed(id) {
-                            if let Err(e) = crate::actions::clipboard::set_text(&time) {
-                                self.report_error_message(
-                                    "launcher",
-                                    format!("Failed to copy time: {e}"),
-                                );
-                            } else if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Copied {time}").into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if let Some(idx_act) = custom_idx {
-                        if ui.button("Edit App").clicked() {
-                            self.editor.open_edit(idx_act, &self.actions[idx_act]);
-                            self.show_editor = true;
-                            ui.close_menu();
-                        }
-                    }
-                    self.pin_result_menu(ui, action);
-                });
-            }
-            ResultContextMenuKind::Snippet => {
-                menu_resp.clone().context_menu(|ui| {
-                    if ui.button("Edit Snippet").clicked() {
-                        self.snippet_dialog.open_edit(&action.label);
-                        ui.close_menu();
-                    }
-                    if ui.button("Remove Snippet").clicked() {
-                        if let Err(e) = remove_snippet(SNIPPETS_FILE, &action.label) {
-                            self.report_error_message(
-                                "launcher",
-                                format!("Failed to remove snippet: {e}"),
-                            );
-                        } else {
-                            *refresh = true;
-                            *set_focus = true;
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Removed snippet {}", action.label).into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if let Some(idx_act) = custom_idx {
-                        if ui.button("Edit App").clicked() {
-                            self.editor.open_edit(idx_act, &self.actions[idx_act]);
-                            self.show_editor = true;
-                            ui.close_menu();
-                        }
-                    }
-                    self.pin_result_menu(ui, action);
-                });
-            }
-            ResultContextMenuKind::Tempfile => {
-                let file_path = action.action.clone();
-                menu_resp.clone().context_menu(|ui| {
-                    if ui.button("Set Alias").clicked() {
-                        self.tempfile_alias_dialog.open(&file_path);
-                        ui.close_menu();
-                    }
-                    if ui.button("Delete File").clicked() {
-                        if let Err(e) =
-                            crate::plugins::tempfile::remove_file(std::path::Path::new(&file_path))
-                        {
-                            self.report_error_message(
-                                "launcher",
-                                format!("Failed to delete file: {e}"),
-                            );
-                        } else {
-                            *refresh = true;
-                            *set_focus = true;
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Removed file {}", action.label).into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if let Some(idx_act) = custom_idx {
-                        if ui.button("Edit App").clicked() {
-                            self.editor.open_edit(idx_act, &self.actions[idx_act]);
-                            self.show_editor = true;
-                            ui.close_menu();
-                        }
-                    }
-                    self.pin_result_menu(ui, action);
-                });
-            }
-            ResultContextMenuKind::Note { slug } => {
-                menu_resp.clone().context_menu(|ui| {
-                    if ui.button("Edit Note").clicked() {
-                        self.open_note_panel(&slug, None);
-                        ui.close_menu();
-                    }
-                    if ui.button("Open in Notepad").clicked() {
-                        match crate::plugins::note::load_notes() {
-                            Ok(notes) => {
-                                if let Some(note) = notes.iter().find(|n| n.slug == slug) {
-                                    if let Err(e) = std::process::Command::new("notepad.exe")
-                                        .arg(&note.path)
-                                        .spawn()
-                                    {
-                                        self.report_error_message("launcher", e.to_string());
-                                    }
-                                } else {
-                                    self.report_error_message(
-                                        "launcher",
-                                        "Note not found".to_string(),
-                                    );
-                                }
-                            }
-                            Err(e) => {
-                                self.report_error_message("launcher", e.to_string());
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if ui.button("Open in Neovim").clicked() {
-                        if self.open_note_in_neovim(
-                            &slug,
-                            crate::plugins::note::load_notes,
-                            |path| spawn_external(path, NoteExternalOpen::Wezterm),
-                        ) {
-                            ui.close_menu();
-                        }
-                    }
-                    if ui.button("Remove Note").clicked() {
-                        self.delete_note(&slug);
-                        *refresh = true;
-                        *set_focus = true;
-                        ui.close_menu();
-                    }
-                    if let Some(idx_act) = custom_idx {
-                        if ui.button("Edit App").clicked() {
-                            self.editor.open_edit(idx_act, &self.actions[idx_act]);
-                            self.show_editor = true;
-                            ui.close_menu();
-                        }
-                    }
-                    self.pin_result_menu(ui, action);
-                });
-            }
-            ResultContextMenuKind::Clipboard { idx, label } => {
-                menu_resp.clone().context_menu(|ui| {
-                    if ui.button("Edit Entry").clicked() {
-                        self.clipboard_dialog.open_edit(idx);
-                        ui.close_menu();
-                    }
-                    if ui.button("Remove Entry").clicked() {
-                        if let Err(e) = crate::plugins::clipboard::remove_entry(
-                            crate::plugins::clipboard::CLIPBOARD_FILE,
-                            idx,
-                        ) {
-                            self.report_error_message(
-                                "launcher",
-                                format!("Failed to remove entry: {e}"),
-                            );
-                        } else {
-                            *refresh = true;
-                            *set_focus = true;
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Removed entry {}", label).into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        ui.close_menu();
-                    }
-                    if let Some(idx_act) = custom_idx {
-                        if ui.button("Edit App").clicked() {
-                            self.editor.open_edit(idx_act, &self.actions[idx_act]);
-                            self.show_editor = true;
-                            ui.close_menu();
-                        }
-                    }
-                    self.pin_result_menu(ui, action);
-                });
-            }
-            ResultContextMenuKind::Todo { idx } => {
-                menu_resp.clone().context_menu(|ui| {
-                    if ui.button("Edit Todo").clicked() {
-                        self.todo_view_dialog.open_edit(idx);
-                        ui.close_menu();
-                    }
-                    if let Some(idx_act) = custom_idx {
-                        if ui.button("Edit App").clicked() {
-                            self.editor.open_edit(idx_act, &self.actions[idx_act]);
-                            self.show_editor = true;
-                            ui.close_menu();
-                        }
-                    }
-                    self.pin_result_menu(ui, action);
-                });
-            }
-            ResultContextMenuKind::Default => {
-                menu_resp.clone().context_menu(|ui| {
-                    if let Some(idx_act) = custom_idx {
-                        if ui.button("Edit App").clicked() {
-                            self.editor.open_edit(idx_act, &self.actions[idx_act]);
-                            self.show_editor = true;
-                            ui.close_menu();
-                        }
-                    }
-                    self.pin_result_menu(ui, action);
-                });
-            }
-        }
-
-        menu_resp
-    }
-
-    fn pin_result_menu(&mut self, ui: &mut egui::Ui, action: &Action) {
-        ui.separator();
-        let pins = history::load_pins(HISTORY_PINS_FILE).unwrap_or_default();
-        let is_pinned = pins.iter().any(|pin| pin.matches_action(action));
-        let pin = HistoryPin {
-            action_id: action.action.clone(),
-            label: action.label.clone(),
-            desc: action.desc.clone(),
-            args: action.args.clone(),
-            query: self.query.clone(),
-            timestamp: chrono::Utc::now().timestamp(),
-        };
-
-        if !is_pinned {
-            if ui.button("Pin current query result").clicked() {
-                match history::upsert_pin(HISTORY_PINS_FILE, &pin) {
-                    Ok(_) => {
-                        if self.enable_toasts {
-                            push_toast(
-                                &mut self.toasts,
-                                Toast {
-                                    text: format!("Pinned {}", action.label).into(),
-                                    kind: ToastKind::Success,
-                                    options: ToastOptions::default()
-                                        .duration_in_seconds(self.toast_duration as f64),
-                                },
-                            );
-                        }
-                    }
-                    Err(e) => {
-                        self.report_error_message("launcher", format!("Failed to pin result: {e}"));
-                    }
-                }
-                ui.close_menu();
-            }
-        } else {
-            if ui.button("Unpin result").clicked() {
-                if let Err(e) =
-                    history::remove_pin(HISTORY_PINS_FILE, &action.action, action.args.as_deref())
-                {
-                    self.report_error_message("launcher", format!("Failed to unpin result: {e}"));
-                } else if self.enable_toasts {
-                    push_toast(
-                        &mut self.toasts,
-                        Toast {
-                            text: format!("Unpinned {}", action.label).into(),
-                            kind: ToastKind::Success,
-                            options: ToastOptions::default()
-                                .duration_in_seconds(self.toast_duration as f64),
-                        },
-                    );
-                }
-                ui.close_menu();
-            }
-            if ui.button("Replace pin with current result").clicked() {
-                match history::upsert_pin(HISTORY_PINS_FILE, &pin) {
-                    Ok(_) => {
-                        if self.enable_toasts {
-                            push_toast(
-                                &mut self.toasts,
-                                Toast {
-                                    text: format!("Updated pin for {}", action.label).into(),
-                                    kind: ToastKind::Success,
-                                    options: ToastOptions::default()
-                                        .duration_in_seconds(self.toast_duration as f64),
-                                },
-                            );
-                        }
-                    }
-                    Err(e) => {
-                        self.report_error_message("launcher", format!("Failed to update pin: {e}"));
-                    }
-                }
-                ui.close_menu();
-            }
-        }
-
-        if ui.button("Recompute pinned results").clicked() {
-            match history::recompute_pins(HISTORY_PINS_FILE, |pin| self.resolve_pin_action(pin)) {
-                Ok(report) => {
-                    if self.enable_toasts {
-                        let text = if report.updated == 0 && report.missing == 0 {
-                            "Pinned results are up to date.".to_string()
-                        } else if report.updated > 0 && report.missing > 0 {
-                            format!(
-                                "Updated {} pinned results ({} missing).",
-                                report.updated, report.missing
-                            )
-                        } else if report.updated > 0 {
-                            format!("Updated {} pinned results.", report.updated)
-                        } else {
-                            format!("{} pinned results missing.", report.missing)
-                        };
-                        push_toast(
-                            &mut self.toasts,
-                            Toast {
-                                text: text.into(),
-                                kind: if report.missing > 0 {
-                                    ToastKind::Warning
-                                } else {
-                                    ToastKind::Success
-                                },
-                                options: ToastOptions::default()
-                                    .duration_in_seconds(self.toast_duration as f64),
-                            },
-                        );
-                    }
-                }
-                Err(e) => {
-                    self.report_error_message("launcher", format!("Failed to recompute pins: {e}"));
-                }
-            }
-            ui.close_menu();
-        }
-    }
-
     pub fn set_last_search_query(&mut self, s: String) {
         self.last_search_query = s;
     }
@@ -2988,799 +1657,6 @@ impl LauncherApp {
 
     pub fn move_cursor_end_flag(&self) -> bool {
         self.move_cursor_end
-    }
-
-    pub fn activate_action(
-        &mut self,
-        a: Action,
-        query_override: Option<String>,
-        source: ActivationSource,
-    ) {
-        if self.maybe_confirm_destructive_action(&a, query_override.clone(), source) {
-            return;
-        }
-        self.activate_action_confirmed(a, query_override, source);
-    }
-
-    fn maybe_confirm_destructive_action(
-        &mut self,
-        a: &Action,
-        query_override: Option<String>,
-        source: ActivationSource,
-    ) -> bool {
-        if !self.require_confirm_destructive {
-            return false;
-        }
-        if let Some(kind) = DestructiveAction::from_action(a) {
-            self.pending_confirm = Some(PendingConfirmAction {
-                action: a.clone(),
-                query_override,
-                source,
-            });
-            self.confirm_modal.open_for_source(kind, Some(source));
-            return true;
-        }
-        false
-    }
-
-    fn activate_action_confirmed(
-        &mut self,
-        a: Action,
-        query_override: Option<String>,
-        source: ActivationSource,
-    ) {
-        if let Some(new_query) = query_override {
-            self.query = new_query;
-            self.last_timer_query =
-                self.query.starts_with("timer list") || self.query.starts_with("alarm list");
-            self.search();
-        }
-        let mut focus_after_launcher = false;
-        if a.action == "launcher:show" {
-            if let Some(query) = a.args.as_ref() {
-                self.query = query.to_string();
-                self.last_timer_query =
-                    query.starts_with("timer list") || query.starts_with("alarm list");
-                self.search();
-                self.move_cursor_end = true;
-                focus_after_launcher = true;
-            }
-        }
-        if self.handle_launcher_action(&a.action) {
-            if focus_after_launcher {
-                self.focus_input();
-            }
-            return;
-        }
-        let current = self.query.clone();
-        let mut refresh = false;
-        let mut set_focus = false;
-        let mut command_changed_query = false;
-        if let Some(new_q) = a.action.strip_prefix("queryexec:") {
-            tracing::debug!("queryexec action via activation: {new_q}");
-            self.query = new_q.to_string();
-            self.last_timer_query =
-                new_q.starts_with("timer list") || new_q.starts_with("alarm list");
-            self.search();
-            self.move_cursor_end = true;
-            if let Some(action) = self.results.first().cloned() {
-                self.activate_action(action, None, source);
-            }
-            return;
-        } else if let Some(new_q) = a.action.strip_prefix("query:") {
-            tracing::debug!("query action via activation: {new_q}");
-            self.query = new_q.to_string();
-            self.last_timer_query =
-                new_q.starts_with("timer list") || new_q.starts_with("alarm list");
-            self.search();
-            self.move_cursor_end = true;
-            self.focus_input();
-            return;
-        } else if a.action == "help:show" {
-            self.help_window.open = true;
-        } else if a.action == "timer:dialog:timer" {
-            self.timer_dialog.open_timer();
-        } else if a.action == "timer:dialog:alarm" {
-            self.timer_dialog.open_alarm();
-        } else if a.action == "calendar:open" || a.action.starts_with("calendar:open:") {
-            let view = a.action.strip_prefix("calendar:open:").unwrap_or("default");
-            let now = chrono::Local::now().naive_local();
-            let mut state =
-                crate::plugins::calendar::load_state(crate::plugins::calendar::CALENDAR_STATE_FILE)
-                    .unwrap_or_default();
-            state.last_opened = Some(now);
-            state.last_viewed_day = Some(now.date());
-            if let Err(err) = crate::plugins::calendar::save_state(
-                crate::plugins::calendar::CALENDAR_STATE_FILE,
-                &state,
-            ) {
-                self.add_error_toast(format!("Calendar state error: {err}"));
-            }
-            if self.dashboard_enabled {
-                self.query.clear();
-                command_changed_query = true;
-                refresh = true;
-                set_focus = true;
-            }
-            self.open_calendar_popover(Some(now.date()));
-            if self.enable_toasts {
-                let label = if view == "default" {
-                    "Opened calendar".to_string()
-                } else {
-                    format!("Opened calendar ({view} view)")
-                };
-                push_toast(
-                    &mut self.toasts,
-                    Toast {
-                        text: label.into(),
-                        kind: ToastKind::Success,
-                        options: ToastOptions::default()
-                            .duration_in_seconds(self.toast_duration as f64),
-                    },
-                );
-            }
-        } else if let Some(reference) = a.action.strip_prefix("calendar:jump:") {
-            let now = chrono::Local::now().naive_local();
-            match crate::plugins::calendar::parse_date_reference(reference, now.date()) {
-                Some(date) => {
-                    let mut state = crate::plugins::calendar::load_state(
-                        crate::plugins::calendar::CALENDAR_STATE_FILE,
-                    )
-                    .unwrap_or_default();
-                    state.last_opened = Some(now);
-                    state.last_viewed_day = Some(date);
-                    if let Err(err) = crate::plugins::calendar::save_state(
-                        crate::plugins::calendar::CALENDAR_STATE_FILE,
-                        &state,
-                    ) {
-                        self.add_error_toast(format!("Calendar state error: {err}"));
-                    }
-                    if self.dashboard_enabled {
-                        self.query.clear();
-                        command_changed_query = true;
-                        refresh = true;
-                        set_focus = true;
-                    }
-                    if self.enable_toasts {
-                        push_toast(
-                            &mut self.toasts,
-                            Toast {
-                                text: format!("Jumped to {}", date.format("%Y-%m-%d")).into(),
-                                kind: ToastKind::Success,
-                                options: ToastOptions::default()
-                                    .duration_in_seconds(self.toast_duration as f64),
-                            },
-                        );
-                    }
-                }
-                None => {
-                    self.add_error_toast(format!("Invalid date reference: {reference}"));
-                }
-            }
-        } else if let Some(input) = a.action.strip_prefix("calendar:add:") {
-            let now = chrono::Local::now().naive_local();
-            match crate::plugins::calendar::parse_calendar_add(input, now) {
-                Ok(request) => match crate::plugins::calendar::add_event(request, now) {
-                    Ok(event) => {
-                        self.dashboard_data_cache.refresh_calendar();
-                        if self.preserve_command {
-                            self.query = "cal add ".into();
-                        } else {
-                            self.query.clear();
-                        }
-                        command_changed_query = true;
-                        refresh = true;
-                        set_focus = true;
-                        if self.enable_toasts {
-                            push_toast(
-                                &mut self.toasts,
-                                Toast {
-                                    text: format!("Added {}", event.title).into(),
-                                    kind: ToastKind::Success,
-                                    options: ToastOptions::default()
-                                        .duration_in_seconds(self.toast_duration as f64),
-                                },
-                            );
-                        }
-                    }
-                    Err(err) => {
-                        self.add_error_toast(format!("Calendar add failed: {err}"));
-                    }
-                },
-                Err(err) => {
-                    self.add_error_toast(err);
-                }
-            }
-        } else if let Some(input) = a.action.strip_prefix("calendar:search:") {
-            match crate::plugins::calendar::parse_calendar_search(input) {
-                Ok(request) => {
-                    let results = crate::plugins::calendar::search_events(&request);
-                    let actions: Vec<Action> = results
-                        .into_iter()
-                        .map(|event| Action {
-                            label: crate::plugins::calendar::format_event_label(&event),
-                            desc: "Calendar".into(),
-                            action: format!("calendar:jump:{}", event.start.format("%Y-%m-%d")),
-                            args: None,
-                        })
-                        .collect();
-                    self.query = format!("cal find {input}");
-                    self.results = actions;
-                    self.selected = None;
-                    self.last_search_query = self.query.clone();
-                    self.last_results_valid = true;
-                    self.update_suggestions();
-                    command_changed_query = true;
-                    set_focus = true;
-                    if self.enable_toasts {
-                        push_toast(
-                            &mut self.toasts,
-                            Toast {
-                                text: format!("Found {} events", self.results.len()).into(),
-                                kind: ToastKind::Info,
-                                options: ToastOptions::default()
-                                    .duration_in_seconds(self.toast_duration as f64),
-                            },
-                        );
-                    }
-                }
-                Err(err) => {
-                    self.add_error_toast(err);
-                }
-            }
-        } else if a.action == "calendar:upcoming" {
-            let now = chrono::Local::now().naive_local();
-            let events = crate::plugins::calendar::CALENDAR_DATA
-                .read()
-                .map(|d| d.clone())
-                .unwrap_or_default();
-            let until = now + chrono::Duration::days(7);
-            let instances = crate::plugins::calendar::expand_instances(&events, now, until, 50);
-            let titles: std::collections::HashMap<_, _> =
-                events.into_iter().map(|e| (e.id, e.title)).collect();
-            self.query = "cal upcoming".into();
-            self.results = instances
-                .into_iter()
-                .map(|instance| {
-                    let title = titles
-                        .get(&instance.source_event_id)
-                        .cloned()
-                        .unwrap_or_else(|| "Calendar event".to_string());
-                    let label = if instance.all_day {
-                        format!("{} ({} all-day)", title, instance.start.format("%Y-%m-%d"))
-                    } else {
-                        format!(
-                            "{} ({} {})",
-                            title,
-                            instance.start.format("%Y-%m-%d"),
-                            instance.start.format("%H:%M")
-                        )
-                    };
-                    Action {
-                        label,
-                        desc: "Calendar".into(),
-                        action: format!("calendar:jump:{}", instance.start.format("%Y-%m-%d")),
-                        args: None,
-                    }
-                })
-                .collect();
-            self.selected = None;
-            self.last_search_query = self.query.clone();
-            self.last_results_valid = true;
-            self.update_suggestions();
-            command_changed_query = true;
-            set_focus = true;
-        } else if let Some(input) = a.action.strip_prefix("calendar:snooze:") {
-            let mut parts = input.split_whitespace();
-            if let (Some(duration_str), Some(event_id)) = (parts.next(), parts.next()) {
-                if let Some(duration) = crate::plugins::calendar::parse_duration_spec(duration_str)
-                {
-                    match crate::plugins::calendar::snooze_event(event_id, duration) {
-                        Ok(true) => {
-                            self.dashboard_data_cache.refresh_calendar();
-                            if self.enable_toasts {
-                                push_toast(
-                                    &mut self.toasts,
-                                    Toast {
-                                        text: format!("Snoozed event {event_id}").into(),
-                                        kind: ToastKind::Success,
-                                        options: ToastOptions::default()
-                                            .duration_in_seconds(self.toast_duration as f64),
-                                    },
-                                );
-                            }
-                        }
-                        Ok(false) => {
-                            self.add_error_toast(format!("Event not found: {event_id}"));
-                        }
-                        Err(err) => {
-                            self.add_error_toast(format!("Snooze failed: {err}"));
-                        }
-                    }
-                } else {
-                    self.add_error_toast("Invalid snooze duration (use 10m, 1h, 2d)");
-                }
-            } else {
-                self.add_error_toast("Provide a duration and event id to snooze");
-            }
-        } else if a.action == "shell:dialog" {
-            self.shell_cmd_dialog.open();
-        } else if a.action == "note:dialog" {
-            self.notes_dialog.open();
-        } else if a.action == "note:graph_dialog" {
-            self.note_graph_dialog.open_with_args(a.args.as_deref());
-        } else if a.action == "note:unused_assets" {
-            self.unused_assets_dialog.open();
-        } else if a.action == "bookmark:dialog" {
-            self.add_bookmark_dialog.open();
-        } else if a.action == "snippet:dialog" {
-            self.snippet_dialog.open();
-        } else if let Some(alias) = a.action.strip_prefix("snippet:edit:") {
-            self.snippet_dialog.open_edit(alias);
-        } else if a.action == "macro:dialog" {
-            self.macro_dialog.open();
-        } else if a.action == "mg:dialog" {
-            self.mouse_gestures_dialog.open();
-        } else if a.action == "mg:dialog:add" {
-            self.mouse_gestures_dialog.open_add();
-        } else if a.action == "mg:dialog:binding" {
-            self.mouse_gestures_dialog.open_binding_editor();
-        } else if a.action == "mg:dialog:focus" {
-            if let Some(args) = a
-                .args
-                .as_deref()
-                .and_then(|raw| serde_json::from_str::<GestureFocusArgs>(raw).ok())
-            {
-                self.mouse_gestures_dialog
-                    .open_focus(&args.label, &args.tokens, args.dir_mode);
-            } else {
-                self.mouse_gestures_dialog.open();
-            }
-        } else if a.action == "mg:dialog:settings" {
-            self.open_mouse_gesture_settings_dialog();
-        } else if a.action == "mg:toggle" {
-            if let Some(args) = a
-                .args
-                .as_deref()
-                .and_then(|raw| serde_json::from_str::<GestureToggleArgs>(raw).ok())
-            {
-                let mut db = load_gestures(GESTURES_FILE).unwrap_or_default();
-                if let Some(gesture) = db.gestures.iter_mut().find(|gesture| {
-                    gesture.label == args.label
-                        && gesture.tokens == args.tokens
-                        && gesture.dir_mode == args.dir_mode
-                }) {
-                    gesture.enabled = args.enabled;
-                    if let Err(err) = save_gestures(GESTURES_FILE, &db) {
-                        self.report_error_message(
-                            "launcher",
-                            format!("Failed to save mouse gestures: {err}"),
-                        );
-                    } else {
-                        self.dashboard_data_cache.refresh_gestures();
-                    }
-                }
-            }
-        } else if let Some(label) = a.action.strip_prefix("fav:dialog:") {
-            if label.is_empty() {
-                self.fav_dialog.open();
-            } else {
-                self.fav_dialog.open_edit(label);
-            }
-        } else if a.action == "todo:dialog" {
-            self.todo_dialog.open();
-        } else if a.action == "todo:view" {
-            self.todo_view_dialog.open();
-        } else if let Some(idx) = a.action.strip_prefix("todo:edit:") {
-            if let Ok(i) = idx.parse::<usize>() {
-                self.todo_view_dialog.open_edit(i);
-            }
-        } else if a.action == "clipboard:dialog" {
-            self.clipboard_dialog.open();
-        } else if let Some(slug) = a.action.strip_prefix("note:open:") {
-            let slug = slug.to_string();
-            self.open_note_panel(&slug, None);
-        } else if let Some(rest) = a.action.strip_prefix("note:new:") {
-            let mut parts = rest.split(':');
-            let slug = parts.next().unwrap_or("").trim();
-            let template = parts.next().map(str::trim);
-            let has_extra_delimiter = parts.next().is_some();
-            let slug_has_whitespace = slug.chars().any(char::is_whitespace);
-            let template_invalid = template
-                .map(|tpl| tpl.is_empty() || tpl.chars().any(char::is_whitespace))
-                .unwrap_or(false);
-            if has_extra_delimiter || slug.is_empty() || slug_has_whitespace || template_invalid {
-                self.report_error_message(
-                    "launcher",
-                    format!("Malformed note action: {}", a.action),
-                );
-                return;
-            }
-            self.open_note_panel(slug, template);
-        } else if a.action == "note:tags" {
-            self.open_note_tags();
-            set_focus = true;
-        } else if let Some(link) = a.action.strip_prefix("note:link:") {
-            self.open_note_link(link);
-        } else if let Some(link_id) = a.action.strip_prefix("link:open:") {
-            if let Ok(parsed) = crate::linking::parse_link_id(link_id) {
-                match parsed.target_type {
-                    crate::linking::LinkTarget::Note => {
-                        self.open_note_panel(&parsed.target_id, None);
-                    }
-                    crate::linking::LinkTarget::Todo => {
-                        self.query = format!("todo links id:{}", parsed.target_id);
-                        self.search();
-                    }
-                    _ => {
-                        self.report_error_message(
-                            "launcher",
-                            format!("Unsupported link target: {}", link_id),
-                        );
-                    }
-                }
-            } else {
-                self.report_error_message("launcher", format!("Invalid link id: {}", link_id));
-            }
-        } else if let Some(slug) = a.action.strip_prefix("note:remove:") {
-            self.delete_note(slug);
-        } else if a.action == "convert:panel" {
-            self.convert_panel.open();
-        } else if a.action == "tempfile:dialog" {
-            self.tempfile_dialog.open();
-        } else if a.action == "settings:dialog" {
-            self.open_settings_dialog();
-        } else if a.action == "dashboard:settings" {
-            let registry = self.dashboard.registry().clone();
-            self.dashboard_editor.open(&self.dashboard_path, &registry);
-            self.show_dashboard_editor = true;
-        } else if a.action == "theme:dialog" {
-            self.open_theme_settings_dialog();
-        } else if a.action == "volume:dialog" {
-            self.volume_dialog.open();
-        } else if a.action == "brightness:dialog" {
-            self.brightness_dialog.open();
-        } else if let Some(n) = a.action.strip_prefix("sysinfo:cpu_list:") {
-            if let Ok(count) = n.parse::<usize>() {
-                self.cpu_list_dialog.open(count);
-            }
-        } else if a.action.starts_with("tab:switch:") {
-            if self.enable_toasts {
-                push_toast(
-                    &mut self.toasts,
-                    Toast {
-                        text: format!("Switching to {}", a.label).into(),
-                        kind: ToastKind::Info,
-                        options: ToastOptions::default()
-                            .duration_in_seconds(self.toast_duration as f64),
-                    },
-                );
-            }
-            let act = a.clone();
-            std::thread::spawn(move || {
-                if let Err(e) = launch_action(&act) {
-                    tracing::error!(?e, "failed to switch tab");
-                }
-            });
-            if a.action != "help:show" {
-                self.record_history_usage(&a, &current, source);
-            }
-        } else if let Some(mode) = a.action.strip_prefix("screenshot:") {
-            use crate::actions::screenshot::Mode as ScreenshotMode;
-            let (mode, clip, tool) = match mode {
-                "window" => (ScreenshotMode::Window, false, MarkupTool::Rectangle),
-                "region" => (ScreenshotMode::Region, false, MarkupTool::Rectangle),
-                "region_markup" => (ScreenshotMode::Region, false, MarkupTool::Pen),
-                "desktop" => (ScreenshotMode::Desktop, false, MarkupTool::Rectangle),
-                "window_clip" => (ScreenshotMode::Window, true, MarkupTool::Rectangle),
-                "region_clip" => (ScreenshotMode::Region, true, MarkupTool::Rectangle),
-                "desktop_clip" => (ScreenshotMode::Desktop, true, MarkupTool::Rectangle),
-                _ => (ScreenshotMode::Desktop, false, MarkupTool::Rectangle),
-            };
-            let screenshot_result =
-                crate::plugins::screenshot::launch_editor(self, mode, clip, tool);
-            if self.handle_screenshot_launch_result(screenshot_result) && a.action != "help:show" {
-                self.record_history_usage(&a, &current, source);
-            }
-        } else if let Err(e) = execute_action(&a) {
-            if a.desc == "Fav" && !a.action.starts_with("fav:") {
-                tracing::error!(?e, fav=%a.label, "failed to run favorite");
-            }
-            self.report_error_message("launcher", format!("Failed: {e}"));
-            self.add_error_toast(format!("Failed: {e}"));
-        } else {
-            if a.desc == "Fav" && !a.action.starts_with("fav:") {
-                tracing::info!(fav=%a.label, command=%a.action, "ran favorite");
-            }
-            if self.enable_toasts && a.action != "recycle:clean" {
-                let msg = if a.action.starts_with("clipboard:") {
-                    format!("Copied {}", a.label)
-                } else {
-                    format!("Launched {}", a.label)
-                };
-                push_toast(
-                    &mut self.toasts,
-                    Toast {
-                        text: msg.into(),
-                        kind: ToastKind::Success,
-                        options: ToastOptions::default()
-                            .duration_in_seconds(self.toast_duration as f64),
-                    },
-                );
-            }
-            if a.action != "help:show" {
-                self.record_history_usage(&a, &current, source);
-            }
-            if a.action == "note:reload" {
-                refresh = true;
-                set_focus = true;
-                if self.enable_toasts {
-                    push_toast(
-                        &mut self.toasts,
-                        Toast {
-                            text: "Reloaded notes".into(),
-                            kind: ToastKind::Success,
-                            options: ToastOptions::default()
-                                .duration_in_seconds(self.toast_duration as f64),
-                        },
-                    );
-                }
-            } else if a.action.starts_with("bookmark:add:") {
-                if self.preserve_command {
-                    self.query = "bm add ".into();
-                } else {
-                    self.query.clear();
-                }
-                command_changed_query = true;
-                refresh = true;
-                set_focus = true;
-            } else if a.action.starts_with("bookmark:remove:") {
-                refresh = true;
-                set_focus = true;
-            } else if a.action.starts_with("folder:add:") {
-                if self.preserve_command {
-                    self.query = "f add ".into();
-                } else {
-                    self.query.clear();
-                }
-                command_changed_query = true;
-                refresh = true;
-                set_focus = true;
-            } else if a.action.starts_with("folder:remove:") {
-                refresh = true;
-                set_focus = true;
-            } else if a.action.starts_with("fav:add:") {
-                refresh = true;
-                set_focus = true;
-            } else if a.action.starts_with("fav:remove:") {
-                refresh = true;
-                set_focus = true;
-            } else if a.action.starts_with("todo:add:") {
-                if self.preserve_command {
-                    self.query = "todo add ".into();
-                } else {
-                    self.query.clear();
-                }
-                command_changed_query = true;
-                refresh = true;
-                set_focus = true;
-                if self.enable_toasts {
-                    if let Some(text) = a
-                        .action
-                        .strip_prefix("todo:add:")
-                        .and_then(|r| r.split('|').next())
-                    {
-                        push_toast(
-                            &mut self.toasts,
-                            Toast {
-                                text: format!("Added todo {text}").into(),
-                                kind: ToastKind::Success,
-                                options: ToastOptions::default()
-                                    .duration_in_seconds(self.toast_duration as f64),
-                            },
-                        );
-                    }
-                }
-            } else if a.action.starts_with("todo:remove:") {
-                refresh = true;
-                set_focus = true;
-                if current.starts_with("note list") {
-                    self.pending_query = Some(current.clone());
-                    command_changed_query = true;
-                }
-                if self.enable_toasts {
-                    let label = a.label.strip_prefix("Remove todo ").unwrap_or(&a.label);
-                    push_toast(
-                        &mut self.toasts,
-                        Toast {
-                            text: format!("Removed todo {label}").into(),
-                            kind: ToastKind::Success,
-                            options: ToastOptions::default()
-                                .duration_in_seconds(self.toast_duration as f64),
-                        },
-                    );
-                }
-            } else if a.action.starts_with("todo:done:") {
-                refresh = true;
-                set_focus = true;
-                self.pending_query = Some(current.clone());
-                command_changed_query = true;
-                if self.enable_toasts {
-                    let label = a
-                        .label
-                        .trim_start_matches("[x] ")
-                        .trim_start_matches("[ ] ");
-                    push_toast(
-                        &mut self.toasts,
-                        Toast {
-                            text: format!("Toggled todo {label}").into(),
-                            kind: ToastKind::Success,
-                            options: ToastOptions::default()
-                                .duration_in_seconds(self.toast_duration as f64),
-                        },
-                    );
-                }
-            } else if a.action.starts_with("todo:pset:") {
-                refresh = true;
-                set_focus = true;
-                if self.enable_toasts {
-                    push_toast(
-                        &mut self.toasts,
-                        Toast {
-                            text: "Updated todo priority".into(),
-                            kind: ToastKind::Success,
-                            options: ToastOptions::default()
-                                .duration_in_seconds(self.toast_duration as f64),
-                        },
-                    );
-                }
-            } else if a.action.starts_with("todo:tag:") {
-                refresh = true;
-                set_focus = true;
-                if self.enable_toasts {
-                    push_toast(
-                        &mut self.toasts,
-                        Toast {
-                            text: "Updated todo tags".into(),
-                            kind: ToastKind::Success,
-                            options: ToastOptions::default()
-                                .duration_in_seconds(self.toast_duration as f64),
-                        },
-                    );
-                }
-            } else if a.action == "todo:clear" {
-                refresh = true;
-                set_focus = true;
-                if self.enable_toasts {
-                    push_toast(
-                        &mut self.toasts,
-                        Toast {
-                            text: "Cleared completed todos".into(),
-                            kind: ToastKind::Success,
-                            options: ToastOptions::default()
-                                .duration_in_seconds(self.toast_duration as f64),
-                        },
-                    );
-                }
-            } else if a.action.starts_with("snippet:remove:") {
-                refresh = true;
-                set_focus = true;
-                if self.enable_toasts {
-                    push_toast(
-                        &mut self.toasts,
-                        Toast {
-                            text: format!("Removed snippet {}", a.label).into(),
-                            kind: ToastKind::Success,
-                            options: ToastOptions::default()
-                                .duration_in_seconds(self.toast_duration as f64),
-                        },
-                    );
-                }
-            } else if a.action.starts_with("tempfile:remove:") {
-                refresh = true;
-                set_focus = true;
-            } else if a.action.starts_with("tempfile:alias:") {
-                refresh = true;
-                set_focus = true;
-            } else if a.action == "tempfile:new" || a.action.starts_with("tempfile:new:") {
-                if self.preserve_command {
-                    self.query = "tmp new ".into();
-                } else {
-                    self.query.clear();
-                }
-                command_changed_query = true;
-                set_focus = true;
-            } else if a.action.starts_with("timer:cancel:") && current.starts_with("timer rm") {
-                refresh = true;
-                set_focus = true;
-            } else if a.action.starts_with("timer:pause:") && current.starts_with("timer pause") {
-                refresh = true;
-                set_focus = true;
-            } else if a.action.starts_with("timer:resume:") && current.starts_with("timer resume") {
-                refresh = true;
-                set_focus = true;
-            } else if a.action.starts_with("timer:start:") && current.starts_with("timer add") {
-                if self.preserve_command {
-                    self.query = "timer add ".into();
-                } else {
-                    self.query.clear();
-                }
-                command_changed_query = true;
-                set_focus = true;
-            }
-            if self.clear_query_after_run && !command_changed_query {
-                self.query.clear();
-                refresh = true;
-                set_focus = true;
-            }
-            if self.hide_after_run
-                && !a.action.starts_with("bookmark:add:")
-                && !a.action.starts_with("bookmark:remove:")
-                && !a.action.starts_with("folder:add:")
-                && !a.action.starts_with("folder:remove:")
-                && !a.action.starts_with("snippet:remove:")
-                && !a.action.starts_with("fav:add:")
-                && !a.action.starts_with("fav:remove:")
-                && !a.action.starts_with("screenshot:")
-                && !a.action.starts_with("calc:")
-                && !a.action.starts_with("todo:done:")
-            {
-                self.visible_flag.store(false, Ordering::SeqCst);
-            }
-        }
-        if refresh {
-            self.last_results_valid = false;
-            self.search();
-        }
-        let _ = command_changed_query;
-        if set_focus {
-            self.focus_input();
-        } else if self.visible_flag.load(Ordering::SeqCst) && !self.any_panel_open() {
-            self.focus_input();
-        }
-    }
-
-    fn record_history_usage(&mut self, action: &Action, query: &str, source: ActivationSource) {
-        let _ = history::append_history(
-            HistoryEntry {
-                query: query.to_string(),
-                query_lc: String::new(),
-                action: action.clone(),
-                source: Some(source.label().to_string()),
-                timestamp: 0,
-            },
-            self.history_limit,
-        );
-        let count = self.usage.entry(action.action.clone()).or_insert(0);
-        *count += 1;
-    }
-
-    fn handle_launcher_action(&mut self, action: &str) -> bool {
-        match action {
-            "launcher:toggle" => {
-                let next = !self.visible_flag.load(Ordering::SeqCst);
-                self.visible_flag.store(next, Ordering::SeqCst);
-                if next {
-                    self.restore_flag.store(true, Ordering::SeqCst);
-                }
-                true
-            }
-            "launcher:show" => {
-                self.visible_flag.store(true, Ordering::SeqCst);
-                self.restore_flag.store(true, Ordering::SeqCst);
-                true
-            }
-            "launcher:hide" => {
-                self.visible_flag.store(false, Ordering::SeqCst);
-                true
-            }
-            "launcher:focus" | "launcher:restore" => {
-                self.visible_flag.store(true, Ordering::SeqCst);
-                self.restore_flag.store(true, Ordering::SeqCst);
-                true
-            }
-            _ => false,
-        }
     }
 
     fn any_panel_open(&self) -> bool {
@@ -4404,717 +2280,6 @@ impl LauncherApp {
         check!(self.show_editor, editor, Panel::Editor);
         check!(self.show_settings, settings, Panel::Settings);
         check!(self.show_plugins, plugins, Panel::Plugins);
-    }
-}
-
-impl eframe::App for LauncherApp {
-    fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
-        use egui::*;
-
-        // tracing::debug!("LauncherApp::update called");
-        if self.enable_toasts {
-            self.toasts.show(ctx);
-        }
-        let frame_time = Duration::from_secs_f32(ctx.input(|i| i.unstable_dt).max(0.0));
-        self.dashboard.update_frame_timing(frame_time);
-        if let Some(pending) = self.pending_query.take() {
-            self.query = pending;
-            self.search();
-            self.focus_input();
-        }
-        self.maybe_run_note_search_debounce();
-        if let (Some(t), Some(_)) = (self.error_time, self.error.as_ref()) {
-            if t.elapsed().as_secs_f32() >= 3.0 {
-                self.error = None;
-                self.error_time = None;
-            }
-        }
-        if self
-            .enabled_capabilities
-            .as_ref()
-            .and_then(|m| m.get("timer"))
-            .map(|c| c.contains(&"completion_dialog".to_string()))
-            .unwrap_or(true)
-        {
-            for msg in crate::plugins::timer::take_finished_messages() {
-                self.completion_dialog.open_message(msg);
-            }
-        }
-        for msg in crate::plugins::macros::take_step_messages() {
-            if self.enable_toasts {
-                push_toast(
-                    &mut self.toasts,
-                    Toast {
-                        text: msg.into(),
-                        kind: ToastKind::Info,
-                        options: ToastOptions::default()
-                            .duration_in_seconds(self.toast_duration as f64),
-                    },
-                );
-            }
-        }
-        for msg in crate::plugins::browser_tabs::take_cache_messages() {
-            if self.enable_toasts {
-                push_toast(
-                    &mut self.toasts,
-                    Toast {
-                        text: msg.into(),
-                        kind: ToastKind::Info,
-                        options: ToastOptions::default()
-                            .duration_in_seconds(self.toast_duration as f64),
-                    },
-                );
-            }
-        }
-        for err in crate::plugins::macros::take_error_messages() {
-            tracing::debug!("{err}");
-        }
-
-        let dropped = ctx.input(|i| i.raw.dropped_files.clone());
-        self.handle_dropped_files(dropped);
-        if let Some(rect) = ctx.input(|i| i.viewport().inner_rect) {
-            self.window_size = (rect.width() as i32, rect.height() as i32);
-        }
-        if let Some(rect) = ctx.input(|i| i.viewport().outer_rect) {
-            self.window_pos = (rect.min.x as i32, rect.min.y as i32);
-        }
-        let do_restore = self.restore_flag.swap(false, Ordering::SeqCst);
-        if self.visible_flag.load(Ordering::SeqCst) && self.help_flag.swap(false, Ordering::SeqCst)
-        {
-            self.help_window.overlay_open = !self.help_window.overlay_open;
-        } else {
-            // reset any queued toggle when window not visible
-            self.help_flag.store(false, Ordering::SeqCst);
-        }
-        if do_restore {
-            tracing::debug!("Restoring window on restore_flag");
-            apply_visibility(
-                true,
-                ctx,
-                self.offscreen_pos,
-                self.follow_mouse,
-                self.static_location_enabled,
-                self.static_pos.map(|(x, y)| (x as f32, y as f32)),
-                self.static_size.map(|(w, h)| (w as f32, h as f32)),
-                (self.window_size.0 as f32, self.window_size.1 as f32),
-            );
-            if let Some(hwnd) = crate::window_manager::get_hwnd(_frame) {
-                crate::window_manager::force_restore_and_foreground(hwnd);
-            }
-        }
-
-        let should_be_visible = self.visible_flag.load(Ordering::SeqCst);
-        let just_became_visible = !self.last_visible && should_be_visible;
-        if self.last_visible != should_be_visible {
-            tracing::debug!("gui thread -> visible: {}", should_be_visible);
-            apply_visibility(
-                should_be_visible,
-                ctx,
-                self.offscreen_pos,
-                self.follow_mouse,
-                self.static_location_enabled,
-                self.static_pos.map(|(x, y)| (x as f32, y as f32)),
-                self.static_size.map(|(w, h)| (w as f32, h as f32)),
-                (self.window_size.0 as f32, self.window_size.1 as f32),
-            );
-            self.last_visible = should_be_visible;
-        }
-
-        TopBottomPanel::top("menu_bar").show(ctx, |ui| {
-            menu::bar(ui, |ui| {
-                ui.menu_button("File", |ui| {
-                    ui.menu_button("Apps", |ui| {
-                        if ui.button("Edit Apps").clicked() {
-                            self.show_editor = !self.show_editor;
-                        }
-                        if ui.button("Edit Plugins").clicked() {
-                            self.show_plugins = !self.show_plugins;
-                        }
-                    });
-                    if ui.button("Close Application").clicked() {
-                        ctx.send_viewport_cmd(egui::ViewportCommand::Close);
-                        self.unregister_all_hotkeys();
-                        self.visible_flag.store(false, Ordering::SeqCst);
-                        #[allow(unused_assignments)]
-                        {
-                            self.last_visible = false;
-                        }
-                        #[cfg(not(test))]
-                        std::process::exit(0);
-                    }
-                });
-                ui.menu_button("Settings", |ui| {
-                    if ui.button("Edit Settings").clicked() {
-                        self.show_settings = !self.show_settings;
-                    }
-                });
-                ui.menu_button("Help", |ui| {
-                    if ui.button("Command List").clicked() {
-                        self.help_window.open = true;
-                    }
-                    if ui.button("Linking Guide (todo/note/cal)").clicked() {
-                        self.help_window.open = true;
-                        self.help_window.filter = "todo note cal @note: @todo:".into();
-                    }
-                    if ui.button("Quick Help Overlay").clicked() {
-                        self.help_window.overlay_open = true;
-                    }
-                    if ui.button("Open Toast Log").clicked() {
-                        if std::fs::OpenOptions::new()
-                            .create(true)
-                            .write(true)
-                            .open(TOAST_LOG_FILE)
-                            .is_err()
-                        {
-                            self.report_error_message("launcher", "Failed to create log");
-                        } else if let Err(e) = open::that(TOAST_LOG_FILE) {
-                            self.report_error_message(
-                                "launcher",
-                                format!("Failed to open log: {e}"),
-                            );
-                        }
-                    }
-                    if ui.button("View Toast Log").clicked() {
-                        self.toast_log_dialog.open();
-                    }
-                });
-                for panel in self.pinned_panels.clone() {
-                    let label = format!("{:?}", panel);
-                    if ui.button(label).clicked() {
-                        if self.panel_stack.last() == Some(&panel) {
-                            self.toggle_pin(panel);
-                        } else {
-                            self.focus_panel(panel);
-                        }
-                    }
-                }
-            });
-        });
-
-        self.process_watch_events();
-
-        let trimmed = self.query.trim().to_string();
-        let use_dashboard = self.should_show_dashboard(trimmed.as_str());
-        self.maybe_refresh_timer_list();
-        self.maybe_refresh_stopwatch_list();
-        if trimmed.eq_ignore_ascii_case("net")
-            && self.last_net_update.elapsed().as_secs_f32() >= self.net_refresh
-        {
-            self.search();
-            self.last_net_update = Instant::now();
-        }
-
-        CentralPanel::default().show(ctx, |ui| {
-            ui.heading("🚀 Multi Lnchr");
-            if self.should_render_inline_error() {
-                if let Some(err) = &self.error {
-                    ui.colored_label(Color32::RED, err);
-                }
-            }
-
-            scale_ui(ui, self.query_scale, |ui| {
-                let input_id = egui::Id::new("query_input");
-
-                if self.move_cursor_end {
-                    if ui.ctx().memory(|m| m.has_focus(input_id)) {
-                        let len = self.query.chars().count();
-                        tracing::debug!("moving cursor to end: {len}");
-                        ui.ctx().data_mut(|data| {
-                            let state = data
-                                .get_persisted_mut_or_default::<egui::widgets::text_edit::TextEditState>(
-                                    input_id,
-                                );
-                            state.cursor.set_char_range(Some(egui::text::CCursorRange::one(
-                                egui::text::CCursor::new(len),
-                            )));
-                        });
-                        crate::window_manager::send_end_key();
-                        self.move_cursor_end = false;
-                        tracing::debug!("move_cursor_end cleared after moving");
-                    } else {
-                        tracing::debug!("cursor not moved - input not focused");
-                    }
-                }
-
-                let input = ui.add(
-                    egui::TextEdit::singleline(&mut self.query)
-                        .id_source(input_id)
-                        .desired_width(f32::INFINITY),
-                );
-                if just_became_visible || self.focus_query {
-                    input.request_focus();
-                    self.focus_query = false;
-                }
-
-                if input.changed() {
-                    self.autocomplete_index = 0;
-                    if Self::is_note_search_query(&self.query) {
-                        self.last_note_search_change = Some(Instant::now());
-                    } else {
-                        self.last_note_search_change = None;
-                        self.search();
-                    }
-                }
-
-                if self.query_autocomplete && !use_dashboard && !self.suggestions.is_empty() {
-                    ui.vertical(|ui| {
-                        for s in &self.suggestions {
-                            ui.colored_label(Color32::GRAY, s);
-                        }
-                    });
-                }
-
-                if ctx.input(|i| i.key_pressed(egui::Key::Escape)) {
-                    if self.any_panel_open() {
-                        if self.close_front_dialog() {
-                            ctx.input_mut(|i| i.consume_key(egui::Modifiers::NONE, egui::Key::Escape));
-                        }
-                    } else {
-                        self.visible_flag.store(false, Ordering::SeqCst);
-                    }
-                }
-
-                if ctx.input(|i| i.modifiers.command && i.key_pressed(egui::Key::W)) {
-                    if self.any_panel_open() {
-                        if self.close_front_dialog() {
-                            ctx.input_mut(|i| i.consume_key(egui::Modifiers::COMMAND, egui::Key::W));
-                        }
-                    }
-                }
-
-                if ctx.input(|i| i.key_pressed(egui::Key::ArrowDown)) {
-                    self.handle_key(egui::Key::ArrowDown);
-                }
-
-                if ctx.input(|i| i.key_pressed(egui::Key::ArrowUp)) {
-                    self.handle_key(egui::Key::ArrowUp);
-                }
-
-                if ctx.input(|i| i.key_pressed(egui::Key::PageDown)) {
-                    self.handle_key(egui::Key::PageDown);
-                }
-
-                if ctx.input(|i| i.key_pressed(egui::Key::PageUp)) {
-                    self.handle_key(egui::Key::PageUp);
-                }
-                if ctx.input(|i| i.key_pressed(egui::Key::ArrowLeft)) {
-                    self.handle_key(egui::Key::ArrowLeft);
-                }
-
-                if ctx.input(|i| i.key_pressed(egui::Key::ArrowRight)) {
-                    self.handle_key(egui::Key::ArrowRight);
-                }
-
-                if ctx.input(|i| i.key_pressed(egui::Key::Num8)) {
-                    self.handle_key(egui::Key::Num8);
-                }
-
-                if ctx.input(|i| i.key_pressed(egui::Key::Num2)) {
-                    self.handle_key(egui::Key::Num2);
-                }
-
-                if ctx.input(|i| i.key_pressed(egui::Key::Num4)) {
-                    self.handle_key(egui::Key::Num4);
-                }
-
-                if ctx.input(|i| i.key_pressed(egui::Key::Num6)) {
-                    self.handle_key(egui::Key::Num6);
-                }
-
-
-                let tab = ctx.input(|i| i.key_pressed(egui::Key::Tab));
-                let enter = ctx.input(|i| i.key_pressed(egui::Key::Enter));
-                let mut accepted_suggestion = false;
-                if tab || (enter && self.selected.is_none()) {
-                    accepted_suggestion = self.accept_suggestion(tab);
-                }
-                if accepted_suggestion {
-                    ctx.input_mut(|i| {
-                        if tab {
-                            i.consume_key(egui::Modifiers::NONE, egui::Key::Tab);
-                        }
-                        if enter {
-                            i.consume_key(egui::Modifiers::NONE, egui::Key::Enter);
-                        }
-                    });
-                }
-
-                let mut launch_idx: Option<usize> = None;
-                if !accepted_suggestion
-                    && enter
-                    && !self.bookmark_alias_dialog.open
-                    && !self.tempfile_alias_dialog.open
-                    && !self.tempfile_dialog.open
-                    && !self.shell_cmd_dialog.open
-                    && !self.notes_dialog.open
-                    && !self.todo_dialog.open
-                    && !self.todo_view_dialog.open
-                    && self.note_panels.is_empty()
-                    && self.image_panels.is_empty()
-                {
-                    launch_idx = self.handle_key(egui::Key::Enter);
-                }
-
-                if let Some(i) = launch_idx {
-                    if let Some(a) = self.results.get(i) {
-                        let a = a.clone();
-                        self.activate_action(a, None, ActivationSource::Enter);
-                    }
-                }
-            });
-
-            if use_dashboard {
-                self.dashboard_data_cache
-                    .flush_refresh_requests(&self.plugins);
-                if !self.suggestions.is_empty() {
-                    self.autocomplete_index = 0;
-                    self.suggestions.clear();
-                }
-                let dashboard_visible = self.visible_flag.load(Ordering::SeqCst);
-                let dashboard_focused = ctx.input(|i| i.viewport().focused).unwrap_or(true);
-                let has_diagnostics_widget = self.has_diagnostics_widget();
-                let show_diagnostics_widget =
-                    self.show_dashboard_diagnostics || has_diagnostics_widget;
-                let diagnostics = if self.show_dashboard_diagnostics || has_diagnostics_widget {
-                    Some(self.dashboard.diagnostics_snapshot())
-                } else {
-                    None
-                };
-                let dash_ctx = DashboardContext {
-                    actions: &self.actions,
-                    actions_by_id: &self.actions_by_id,
-                    usage: &self.usage,
-                    plugins: &self.plugins,
-                    enabled_plugins: self.enabled_plugins.as_ref(),
-                    default_location: self.dashboard_default_location.as_deref(),
-                    data_cache: &self.dashboard_data_cache,
-                    actions_version: crate::actions::actions_version(),
-                    fav_version: crate::plugins::fav::fav_version(),
-                    notes_version: crate::plugins::note::note_version(),
-                    todo_version: crate::plugins::todo::todo_version(),
-                    calendar_version: crate::plugins::calendar::calendar_version(),
-                    clipboard_version: crate::plugins::clipboard::clipboard_version(),
-                    snippets_version: crate::plugins::snippets::snippets_version(),
-                    dashboard_visible,
-                    dashboard_focused,
-                    reduce_dashboard_work_when_unfocused: self
-                        .reduce_dashboard_work_when_unfocused,
-                    diagnostics,
-                    show_diagnostics_widget,
-                };
-                ctx.request_repaint_after(Duration::from_millis(250));
-                if let Some(action) = self.dashboard.ui(ui, &dash_ctx, WidgetActivation::Click) {
-                    self.activate_action(action.action, action.query_override, ActivationSource::Dashboard);
-                }
-            } else {
-                let area_height = ui.available_height();
-                ScrollArea::vertical()
-                    .max_height(area_height)
-                    .show(ui, |ui| {
-                        scale_ui(ui, self.list_scale, |ui| {
-                            let mut refresh = false;
-                            let mut set_focus = false;
-                            let show_full = self
-                                .enabled_capabilities
-                                .as_ref()
-                                .and_then(|m| m.get("folders"))
-                                .map(|caps| caps.contains(&"show_full_path".to_string()))
-                                .unwrap_or(false);
-                            if self.resolved_grid_layout {
-                                let cols = self.query_results_layout.cols.max(1);
-                                let col_width = ((ui.available_width()
-                                    - ((cols.saturating_sub(1)) as f32 * 8.0))
-                                    / cols as f32)
-                                    .max(160.0);
-                                egui::Grid::new("query_results_grid")
-                                    .num_columns(cols)
-                                    .spacing([8.0, 6.0])
-                                    .show(ui, |ui| {
-                                        for idx in 0..self.results.len() {
-                                            let action = self.results[idx].clone();
-                                            let text = format!("{}\n{}", action.label, action.desc);
-                                            let resp = ui.add_sized(
-                                                [col_width, 44.0],
-                                                egui::SelectableLabel::new(
-                                                    self.selected == Some(idx),
-                                                    text,
-                                                ),
-                                            );
-                                            let menu_resp = self.attach_result_context_menu(
-                                                &action,
-                                                resp,
-                                                &mut refresh,
-                                                &mut set_focus,
-                                            );
-                                            if self.selected == Some(idx) {
-                                                menu_resp.scroll_to_me(Some(egui::Align::Center));
-                                            }
-                                            if menu_resp.clicked() {
-                                                self.selected = Some(idx);
-                                                self.activate_action(
-                                                    action,
-                                                    None,
-                                                    ActivationSource::Click,
-                                                );
-                                            }
-                                            if (idx + 1) % cols == 0 {
-                                                ui.end_row();
-                                            }
-                                        }
-                                    });
-                            } else {
-                                for idx in 0..self.results.len() {
-                                    let a = self.results[idx].clone();
-                                    let aliased =
-                                        self.folder_aliases.get(&a.action).and_then(|v| v.as_ref());
-                                    let show_path = show_full || aliased.is_none();
-                                    let text = if show_path {
-                                        format!("{} : {}", a.label, a.desc)
-                                    } else {
-                                        a.label.clone()
-                                    };
-                                    let resp = ui.add_sized(
-                                        [ui.available_width(), 0.0],
-                                        egui::SelectableLabel::new(self.selected == Some(idx), text),
-                                    );
-                                    let tooltip = if a.desc == "Timer" && a.action.starts_with("timer:show:") {
-                                        if let Ok(id) = a.action[11..].parse::<u64>() {
-                                            if let Some(ts) = crate::plugins::timer::timer_start_ts(id) {
-                                                format!("Started {}", crate::plugins::timer::format_ts(ts))
-                                            } else {
-                                                a.action.clone()
-                                            }
-                                        } else {
-                                            a.action.clone()
-                                        }
-                                    } else {
-                                        a.action.clone()
-                                    };
-                                    let menu_resp =
-                                        self.attach_result_context_menu(&a, resp.on_hover_text(tooltip), &mut refresh, &mut set_focus);
-                                    if self.selected == Some(idx) {
-                                        menu_resp.scroll_to_me(Some(egui::Align::Center));
-                                    }
-                                    if menu_resp.clicked() {
-                                        self.selected = Some(idx);
-                                        self.activate_action(a.clone(), None, ActivationSource::Click);
-                                    }
-                                }
-                            }
-                            if refresh {
-                                self.last_results_valid = false;
-                                self.search();
-                            }
-                            if set_focus {
-                                self.focus_input();
-                            } else if self.visible_flag.load(Ordering::SeqCst) && !self.any_panel_open()
-                            {
-                                self.focus_input();
-                            }
-                        });
-                    });
-            }
-        });
-        let show_editor = self.show_editor;
-        if show_editor {
-            let mut editor = std::mem::take(&mut self.editor);
-            editor.ui(ctx, self);
-            self.editor = editor;
-        }
-        let show_settings = self.show_settings;
-        if show_settings {
-            let mut ed = std::mem::take(&mut self.settings_editor);
-            ed.ui(ctx, self);
-            self.settings_editor = ed;
-        }
-        let show_plugin = self.show_plugins;
-        if show_plugin {
-            let mut ed = std::mem::take(&mut self.plugin_editor);
-            ed.ui(ctx, self);
-            self.plugin_editor = ed;
-        }
-        if self.show_dashboard_editor && !self.dashboard_editor.open {
-            let registry = self.dashboard.registry().clone();
-            self.dashboard_editor.open(&self.dashboard_path, &registry);
-        }
-        if self.show_dashboard_editor {
-            let registry = self.dashboard.registry().clone();
-            let mut dlg = std::mem::take(&mut self.dashboard_editor);
-            let plugin_infos = self.plugins.plugin_infos();
-            let plugin_commands = self.plugins.commands();
-            let settings_ctx = WidgetSettingsContext {
-                plugins: Some(&self.plugins),
-                plugin_infos: Some(&plugin_infos),
-                plugin_commands: Some(&plugin_commands),
-                actions: Some(self.actions.as_slice()),
-                usage: Some(&self.usage),
-                default_location: self.dashboard_default_location.as_deref(),
-                enabled_plugins: self.enabled_plugins.as_ref(),
-            };
-            let reload = dlg.ui(
-                ctx,
-                &registry,
-                settings_ctx,
-                self.require_confirm_destructive,
-            );
-            self.show_dashboard_editor = dlg.open;
-            self.dashboard_editor = dlg;
-            if reload {
-                self.dashboard.reload();
-            }
-        }
-        let mut dlg = std::mem::take(&mut self.alias_dialog);
-        dlg.ui(ctx, self);
-        self.alias_dialog = dlg;
-        let mut bm_dlg = std::mem::take(&mut self.bookmark_alias_dialog);
-        bm_dlg.ui(ctx, self);
-        self.bookmark_alias_dialog = bm_dlg;
-        let mut tf_dlg = std::mem::take(&mut self.tempfile_alias_dialog);
-        tf_dlg.ui(ctx, self);
-        self.tempfile_alias_dialog = tf_dlg;
-        let mut create_tf = std::mem::take(&mut self.tempfile_dialog);
-        create_tf.ui(ctx, self);
-        self.tempfile_dialog = create_tf;
-        let mut add_bm_dlg = std::mem::take(&mut self.add_bookmark_dialog);
-        add_bm_dlg.ui(ctx, self);
-        self.add_bookmark_dialog = add_bm_dlg;
-        let mut help = std::mem::take(&mut self.help_window);
-        help.ui(ctx, self);
-        self.help_window = help;
-        let mut timer_dlg = std::mem::take(&mut self.timer_dialog);
-        timer_dlg.ui(ctx, self);
-        self.timer_dialog = timer_dlg;
-        let mut comp = std::mem::take(&mut self.completion_dialog);
-        comp.ui(ctx);
-        self.completion_dialog = comp;
-        let mut shell_dlg = std::mem::take(&mut self.shell_cmd_dialog);
-        shell_dlg.ui(ctx, self);
-        self.shell_cmd_dialog = shell_dlg;
-        let mut snip_dlg = std::mem::take(&mut self.snippet_dialog);
-        snip_dlg.ui(ctx, self);
-        self.snippet_dialog = snip_dlg;
-        let mut macro_dlg = std::mem::take(&mut self.macro_dialog);
-        macro_dlg.ui(ctx, self);
-        self.macro_dialog = macro_dlg;
-        let mut mg_dlg = std::mem::take(&mut self.mouse_gestures_dialog);
-        mg_dlg.ui(ctx, self);
-        self.mouse_gestures_dialog = mg_dlg;
-        let mut mg_settings_dlg = std::mem::take(&mut self.mouse_gesture_settings_dialog);
-        mg_settings_dlg.ui(ctx, self);
-        self.mouse_gesture_settings_dialog = mg_settings_dlg;
-        let mut theme_state = std::mem::take(&mut self.theme_settings_dialog);
-        let mut theme_open = self.theme_settings_dialog_open;
-        crate::gui::theme_settings_dialog::ui(ctx, self, &mut theme_open, &mut theme_state);
-        self.theme_settings_dialog_open = theme_open;
-        self.theme_settings_dialog = theme_state;
-        let mut fav_dlg = std::mem::take(&mut self.fav_dialog);
-        fav_dlg.ui(ctx, self);
-        self.fav_dialog = fav_dlg;
-        let mut notes_dlg = std::mem::take(&mut self.notes_dialog);
-        notes_dlg.ui(ctx, self);
-        self.notes_dialog = notes_dlg;
-        let mut graph_dlg = std::mem::take(&mut self.note_graph_dialog);
-        let data_cache: *const DashboardDataCache = &self.dashboard_data_cache;
-        // SAFETY: `data_cache` points to a stable field on `self` for this call. The dialog
-        // only reads through `&DashboardDataCache` while `self` is mutably borrowed for app
-        // actions; no mutation of `dashboard_data_cache` occurs here.
-        let data_cache = unsafe { &*data_cache };
-        graph_dlg.ui(ctx, self, data_cache, crate::plugins::note::note_version());
-        self.note_graph_dialog = graph_dlg;
-        let mut assets_dlg = std::mem::take(&mut self.unused_assets_dialog);
-        assets_dlg.ui(ctx, self);
-        self.unused_assets_dialog = assets_dlg;
-        let mut i = 0;
-        while i < self.note_panels.len() {
-            let mut panel = self.note_panels.remove(i);
-            panel.ui(ctx, self);
-            if panel.open {
-                self.note_panels.insert(i, panel);
-                i += 1;
-            }
-        }
-        let mut i = 0;
-        while i < self.image_panels.len() {
-            let mut panel = self.image_panels.remove(i);
-            panel.ui(ctx);
-            if panel.open {
-                self.image_panels.insert(i, panel);
-                i += 1;
-            }
-        }
-        let mut i = 0;
-        while i < self.screenshot_editors.len() {
-            let mut editor = self.screenshot_editors.remove(i);
-            editor.ui(ctx, self);
-            if editor.open {
-                self.screenshot_editors.insert(i, editor);
-                i += 1;
-            }
-        }
-        let mut todo_dlg = std::mem::take(&mut self.todo_dialog);
-        todo_dlg.ui(ctx, self);
-        self.todo_dialog = todo_dlg;
-        let mut todo_view = std::mem::take(&mut self.todo_view_dialog);
-        todo_view.ui(ctx, self);
-        self.todo_view_dialog = todo_view;
-        let mut cb_dlg = std::mem::take(&mut self.clipboard_dialog);
-        cb_dlg.ui(ctx, self);
-        self.clipboard_dialog = cb_dlg;
-        let mut conv_panel = std::mem::take(&mut self.convert_panel);
-        conv_panel.ui(ctx, self);
-        self.convert_panel = conv_panel;
-        let mut vol_dlg = std::mem::take(&mut self.volume_dialog);
-        vol_dlg.ui(ctx, self);
-        self.volume_dialog = vol_dlg;
-        let mut bright_dlg = std::mem::take(&mut self.brightness_dialog);
-        bright_dlg.ui(ctx, self);
-        self.brightness_dialog = bright_dlg;
-        let mut cpu_dlg = std::mem::take(&mut self.cpu_list_dialog);
-        cpu_dlg.ui(ctx, self);
-        self.cpu_list_dialog = cpu_dlg;
-        let mut toast_dlg = std::mem::take(&mut self.toast_log_dialog);
-        toast_dlg.ui(ctx, self);
-        self.toast_log_dialog = toast_dlg;
-        let mut calendar_popover = std::mem::take(&mut self.calendar_popover);
-        calendar_popover.ui(ctx, self);
-        self.calendar_popover = calendar_popover;
-        let mut calendar_editor = std::mem::take(&mut self.calendar_event_editor);
-        calendar_editor.ui(ctx, self);
-        self.calendar_event_editor = calendar_editor;
-        let mut calendar_details = std::mem::take(&mut self.calendar_event_details);
-        calendar_details.ui(ctx, self);
-        self.calendar_event_details = calendar_details;
-        match self.confirm_modal.ui(ctx) {
-            ConfirmationResult::Confirmed => {
-                if let Some(pending) = self.pending_confirm.take() {
-                    self.activate_action_confirmed(
-                        pending.action,
-                        pending.query_override,
-                        pending.source,
-                    );
-                }
-            }
-            ConfirmationResult::Cancelled => {
-                self.pending_confirm = None;
-            }
-            ConfirmationResult::None => {}
-        }
-        self.enforce_pinned();
-        self.update_panel_stack();
-    }
-
-    fn on_exit(&mut self, _gl: Option<&eframe::glow::Context>) {
-        self.unregister_all_hotkeys();
-        self.visible_flag.store(false, Ordering::SeqCst);
-        self.last_visible = false;
-        if let Ok(mut settings) = crate::settings::Settings::load(&self.settings_path) {
-            settings.window_size = Some(self.window_size);
-            settings.pinned_panels = self.pinned_panels.clone();
-            let _ = settings.save(&self.settings_path);
-        }
-        let _ = usage::save_usage(USAGE_FILE, &self.usage);
-        #[cfg(not(test))]
-        std::process::exit(0);
     }
 }
