+++ tests/plugin_exact_match.rs
@@ -131,16 +131,14 @@ fn note_today_returns_resolved_note_action_when_exact_match_enabled() {
     app.search();
 
     assert!(!app.results.is_empty());
-    assert!(
-        app.results
-            .iter()
-            .any(|result| result.action.starts_with("note:new:"))
-    );
-    assert!(
-        app.results
-            .iter()
-            .all(|result| !result.action.starts_with("query:"))
-    );
+    assert!(app
+        .results
+        .iter()
+        .any(|result| result.action.starts_with("note:new:")));
+    assert!(app
+        .results
+        .iter()
+        .all(|result| !result.action.starts_with("query:")));
 }
 
 #[test]
@@ -161,16 +159,14 @@ fn note_search_matches_note_content_when_exact_match_enabled() {
     app.search();
 
     assert!(!app.results.is_empty());
-    assert!(
-        app.results
-            .iter()
-            .any(|result| result.action == "note:open:beta")
-    );
-    assert!(
-        app.results
-            .iter()
-            .all(|result| !result.action.starts_with("query:"))
-    );
+    assert!(app
+        .results
+        .iter()
+        .any(|result| result.action == "note:open:beta"));
+    assert!(app
+        .results
+        .iter()
+        .all(|result| !result.action.starts_with("query:")));
 }
 
 #[test]
@@ -189,14 +185,12 @@ fn note_new_generates_slugged_action_when_exact_match_enabled() {
     app.search();
 
     assert!(!app.results.is_empty());
-    assert!(
-        app.results
-            .iter()
-            .any(|result| result.action == "note:new:hello-world")
-    );
-    assert!(
-        app.results
-            .iter()
-            .all(|result| !result.action.starts_with("query:"))
-    );
+    assert!(app
+        .results
+        .iter()
+        .any(|result| result.action == "note:new:hello-world"));
+    assert!(app
+        .results
+        .iter()
+        .all(|result| !result.action.starts_with("query:")));
 }
